### 2024-09-27 #7297 @editorialbot
[PRE REVIEW]: sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables
**Submitting author:** @JiangmeiRubyXiong (<a href="http://orcid.org/0000-0002-7481-765X">Jiangmei Xiong</a>)
**Repository:** https://github.com/dragontaoran/sleev
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.3
