### 2024-09-27 #7297 @JiangmeiRubyXiong
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/11071810127?check_suite_focus=true). Author (Jiangmei Xiong) is missing affiliation.

Affiliation updated and issue resolved.
