### 2024-09-27 #7297 @JiangmeiRubyXiong
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - None
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: An empirical study for impacts of measurement erro...
> - No DOI given, and none found for title: Efficient odds ratio estimation under two-phase sa...
> - No DOI given, and none found for title: Spline functions: basic theory
> - No DOI given, and none found for title: On profile likelihood
>
> ❌ MISSING DOIs
>
> - 10.1002/sim.8799 may be a valid DOI for title: Efficient semiparametric inference for two-phase s...
> - 10.1080/01621459.2017.1295864 may be a valid DOI for title: Efficient semiparametric inference under two-phase...
> - 10.1214/20-aoas1343 may be a valid DOI for title: Accounting for dependent errors in predictors and ...
>
> ❌ INVALID DOIs
>
> - None
> ```

DOIs are now added to the paper.bib
