### 2024-10-01 #7297 @jbytecode
@JiangmeiRubyXiong - Hi, thank you for submitting your manuscript and software to JOSS. I am the handling editor of this submission.

While I study your work and try to find a list of suitable reviewers, please take a look at the issue below:

In the manuscript the citations `(@tao2021efficient; @lotspeich2022efficient)` look a little bit akward. Could you please read the documentations

- https://joss.readthedocs.io/en/latest/paper.html#citations
- https://pandoc.org/MANUAL#extension-citations

and correct the citation style of your citations in your paper?

Thank you in advance.
