### 2024-10-01 #7297 @JiangmeiRubyXiong
@jbytecode All citations and author DOIs are now updated! Please let me know if there is anything else I can help.
