### 2024-10-03 #7297 @alemermartinez
I agree to review this submission.

________________________________
De: Mehmet Hakan Satman ***@***.***>
Enviado: martes, 1 de octubre de 2024 13:46
Para: openjournals/joss-reviews ***@***.***>
Cc: Alejandra Martinez ***@***.***>; Mention ***@***.***>
Asunto: Re: [openjournals/joss-reviews] [PRE REVIEW]: sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables (Issue #7297)

👋👋👋 Dear @alemermartinez<https://github.com/alemermartinez> , @aalfons<https://github.com/aalfons> 👋👋👋

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables<https://github.com/openjournals/joss-reviews/issues/7297#top>

You can find more information at the top of this Github issue (#7297<https://github.com/openjournals/joss-reviews/issues/7297#top>).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be about 25 check items for each single reviewer.

Thank you in advance!
