### 2024-10-05 #7297 @tobiasschoch
Dear Mehmet,

I won't be able to do the review until the beginning of next year.

We are in the process of introducing a new curriculum at my university. I'm
co-lead for the re-organization of all statistics courses, which takes up a
lot of time, so I've suspended all research-related activities,
unfortunately.

Kind regards,
Tobias

Am Do., 3. Okt. 2024 um 07:08 Uhr schrieb Mehmet Hakan Satman <
***@***.***>:

> 👋👋👋 Dear @ejikeugba <https://github.com/ejikeugba> , @tobiasschoch
> <https://github.com/tobiasschoch> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> *sleev: An R Package for Semiparametric Likelihood Estimation with Errors
> in Variables <https://github.com/openjournals/joss-reviews/issues/7297#top>*
>
> You can find more information at the top of this Github issue (#7297
> <https://github.com/openjournals/joss-reviews/issues/7297#top>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be about 25 check items for each single reviewer.
>
> Thank you in advance!
>
