### 2024-10-06 #7297 @jbytecode
👋👋👋 Dear @msalibian 👋👋👋

**[sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables](https://github.com/openjournals/joss-reviews/issues/7297#top)**

You can find more information at the top of this Github issue ([#7297](https://github.com/openjournals/joss-reviews/issues/7297#top)).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be about 25 check items for each single reviewer.

Thank you in advance!
