### 2024-10-07 #7320 @editorialbot
[REVIEW]: sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables
**Submitting author:** @JiangmeiRubyXiong (<a href="http://orcid.org/0000-0002-7481-765X">Jiangmei Xiong</a>)
**Repository:** https://github.com/dragontaoran/sleev
**Branch with paper.md** (empty if default branch): jossSubmission
**Version:** v1.1.4
