### 2024-10-22 #7320 @alemermartinez
Hi @jbytecode  I would like to ask for your guidance on how to communicate some issues I observed while reviewing the paper.

For instance, I believe that requiring users to manually select the B-splines basis, without providing a default option for the Bspline argument, makes the package less "user-friendly" than it could be. Including a default option for the univariate case (and for the bivariate setting, if possible) would be beneficial. In situations where computing k-fold cross-validation is computationally expensive, using a fixed number of B-splines as a starting point could be an effective approach. Additionally, it would be important to include some relevant methods to enhance usability.

The term “robust” has been used in the paper. I would appreciate some clarification on what “robust” refers to in this context.

Regarding related packages, are there any other commonly-used packages that estimate under these models, aside from the two developed by the author’s colleagues?

Additionally, I would suggest including system times in order to compare the different B-spline methods presented in the vignette.

I also found some typos, particularly in the mathematical equations. Some of these equations also extend beyond the right margin of the PDF file, affecting readability.
