### 2024-10-23 #7320 @jbytecode
@alemermartinez - Thank you for your review and suggestions. I think they're very clear and I hope @JiangmeiRubyXiong can handle most of them. You can also open pull requests that include your corrections on the manuscript and/or open issues that address your suggestions. Writing down your suggestions here is also a convenient way of reviewing. Thank you!

@JiangmeiRubyXiong - Could you please consider the suggestions above and update your status? Thank you in advance.
