### 2024-10-23 #7320 @alemermartinez
Thanks @jbytecode Below are the additional issues I found while reviewing the paper and vignette.

From the paper:
* Line 26: The word “and” is missing.
* Lines: 45 and 54. The letter “beta” is missing in the PDF compilation.
* Line 49: The symbol for independence might not be clear to the reader. I suggest replacing it with the word “independent”.
* After equation (1): There is no mention in the main paper of the probability term that was left out.
* Line 62: The equation exceeds the page margins and is unreadable.
* Equation (2). The variable “s_n” is lowercase but in the sum it appears in uppercase. Additionally, an “i” appears instead of a “k” and some “k”’s are missing.
* Lines 115 and 117: The formulas exceed the right margin and can be broken before “degree=3”.
* Lines 138 and 139: The formulas exceed the right margin as well.
* Lines 143: The output is missing.
* Table 2: This table is not referenced in the text.

From the vignette:
* The reference to Schumaker should be 2007, as this is the third edition (as mentioned in the paper), but it is cited as 1981 here.
* Before Equation (1), the regression model is stated as Y=alpha + X * beta + epsilon. However, in the first paragraph, it states alpha + beta^T * X + epsilon. Is X the same in both cases? If so, a transpose might be missing.
* Equation (3): There are two logarithms in the second sum.
