+++ vignettes/sleev_vignette.qmd
@@ -64,7 +64,6 @@ editor:
   markdown: 
     wrap: 72
 execute: 
-  cache: true
   warning: false
 ---
 
@@ -131,13 +130,13 @@ implement, as they involve approximating nuisance conditional densities
 using B-splines [@schumaker1981] and then maximizing the semiparametric
 likelihood via a sophisticated EM algorithm [@tao2017efficient].
 
-In this paper, we introduce the sleev package, which computes SMLEs for
+In this paper, we introduce the `sleev` package, which computes SMLEs for
 linear and logistic regressions using partially-validated, error-prone
-data. The sleev package incorporates error-prone data on all records
+data. The `sleev` package incorporates error-prone data on all records
 plus validated data on a subset of records to obtain efficient and
 robust estimates of regression parameters in a user-friendly manner.
 This paper describes the SMLE method (Sections 2 and 3) and demonstrates
-the features of the sleev package and the application of functions in
+the features of the `sleev` package and the application of functions in
 the package with a detailed illustration using simulated HIV data
 (Section 4).
 
@@ -217,7 +216,7 @@ note that the $p_{kj}$'s need to satisfy the constraints
 $\sum_{k=1}^mp_{kj}=1$ and $p_{kj} \geq 0$ since they approximate
 conditional densities. The log-likelihood in expression (@eq-loglik) is
 now approximated by $$
-\sum^n_{i=1}V_i\{\log P_{\pmb{\theta}}(Y_i|\pmb{X}_i)+\sum_{k=1}^mI[(W_i=w_k, \pmb{U}_i=\pmb{u}_k)\sum_{j=1}^{s_n}B_j^q(\pmb{X}_i^{\pmb{*}})\log p_{kj}]\}$$
+\sum^n_{i=1}V_i\{\log P_{\pmb{\theta}}(Y_i|\pmb{X}_i)+\sum_{k=1}^mI(W_i=w_k, \pmb{U}_i=\pmb{u}_k)\sum_{j=1}^{s_n}B_j^q(\pmb{X}_i^{\pmb{*}})\log p_{kj}\}$$
 $$    +\sum^n_{i=1}(1-V_i)\log\left\{\sum_{k=1}^m P_{\pmb{\theta}}(Y^*_i-w_k|\pmb{X}_i^{\pmb{*}}-\pmb{u}_k)\sum_{j=1}^{s_n}B_j^q(\pmb{X}_i^{\pmb{*}})p_{kj}\right\}.  
 $$ {#eq-loglikApprox} The maximization of expression (@eq-loglikApprox)
 is carried out through an EM algorithm to find the SMLEs
@@ -275,9 +274,9 @@ properties, can be found in @lotspeich2022efficient.
 
 # 4. Case study with mock data
 
-## 4.1 Overview and installation of the sleev R package
+## 4.1 Overview and installation of the `sleev` R package
 
-The sleev package provides a user-friendly way to obtain the SMLEs and
+The `sleev` package provides a user-friendly way to obtain the SMLEs and
 their standard errors. The package can be installed through CRAN:
 
 ```         
@@ -289,7 +288,7 @@ library(sleev)
 library(sleev)
 ```
 
-The sleev package includes two main functions: `linear2ph()` and
+The `sleev` package includes two main functions: `linear2ph()` and
 `logistic2ph()`, to fit linear and logistic regressions, respectively,
 under two-phase sampling with an error-prone outcome and covariates. The
 input arguments are similar for the two functions and listed in
@@ -298,7 +297,7 @@ arguments for error-prone and error-free outcome and covariates, the
 user needs to specify the B-spline matrix $B_j^q(\pmb{X}_i^{\pmb{*}})$
 to be used in the estimation of the error densities.
 
[…]
