+++ vignettes/sleev_vignette.qmd
@@ -486,7 +486,7 @@ res_linear <- linear2ph(Y_unval="CD4_unval_sq10", Y="CD4_val_sq10",
                         hn_scale = 1, noSE = FALSE, TOL = 1e-04, 
                         MAX_ITER = 1000, verbose = FALSE)
 paste0("Run time: ", round(difftime(Sys.time(), start.time, 
-                                    units = "mins"), 3), " min")
+                                    units = "secs"), 3), " sec")
 ```
 
 The fitted SMLEs are stored in a `list` object, denoted by `res_linear`
@@ -583,7 +583,7 @@ set.seed(1)
 # different B-spline sizes
 sns <- c(15, 20, 25, 30, 35, 40) 
 # vector to hold mean log-likelihood and run time for each sn
-pred_loglike.1 <- run.time.mins <- rep(NA, length(sns)) 
+pred_loglike.1 <- run.time.secs <- rep(NA, length(sns)) 
 # get number of rows of the dataset
 n <- nrow(mock.vccc) 
 # specify number of folds in the cross validation
@@ -613,13 +613,13 @@ for (i in 1:length(sns)) {
                         nfolds=k, MAX_ITER = 2000, TOL = 1e-04, 
                         verbose = FALSE)
   # save run time
-  run.time.mins[i] <- difftime(Sys.time(), start.time, units = "mins")
+  run.time.secs[i] <- difftime(Sys.time(), start.time, units = "secs")
   # save mean log-likelihood result                  
   pred_loglike.1[i] <- res.1$avg_pred_loglik
 }
 ```
 
-The average predicted log-likelihoods for the different $s_n$ considered
+The average predicted log-likelihoods and run time for the different $s_n$ considered
 are:
 
 ```{r include=FALSE}
@@ -627,7 +627,7 @@ maxns <- sns[which.max(pred_loglike.1)]
 ```
 
 ```{r}
-out <- data.frame(sns, pred_loglike.1, run.time.mins)
+out <- data.frame(sns, pred_loglike.1, run.time.secs)
 options(digits = 6)
 print(out, row.names = FALSE)
 ```
@@ -669,7 +669,7 @@ fit.sn.35 <- linear2ph(Y="CD4_val_sq10", Y_unval="CD4_unval_sq10",
                        hn_scale = 1, noSE = FALSE, TOL = 1e-04, 
                        MAX_ITER = 1000, verbose = FALSE)
 paste0("Run time: ", round(difftime(Sys.time(), start.time, 
-                                    units = "mins"), 3), " min")
+                                    units = "secs"), 3), " sec")
 # compare the coefficients to those from Section 4.3.2
 res_linear$coefficients
 fit.sn.35$coefficients
@@ -770,7 +770,7 @@ res_linear = linear2ph(Y="CD4_val_sq10", Y_unval="CD4_unval_sq10",
                        data=data, hn_scale = 1/4, noSE = FALSE,
                        TOL = 1e-04, MAX_ITER = 1000,verbose = FALSE)
 paste0("Run time: ", round(difftime(Sys.time(), start.time, 
-                                    units = "mins"), 3), " min")
+                                    units = "secs"), 3), " sec")
 res_linear$coefficients
 ```
 
@@ -819,7 +819,7 @@ res_linear = linear2ph(Y="CD4_val_sq10", Y_unval="CD4_unval_sq10",
                        data=data, hn_scale = 1/4, noSE = FALSE, 
                        TOL = 1e-04, MAX_ITER = 1000, verbose = FALSE)
 paste0("Run time: ", round(difftime(Sys.time(), start.time, 
-                                    units = "mins"), 3), " min")
+                                    units = "secs"), 3), " sec")
 res_linear$coefficients
 ```
 
[…]
