### 2024-12-10 #7320 @JiangmeiRubyXiong
Hi @alemermartinez  @jbytecode , thank you for your patience. Here is our response towards the reviewer comments so far. Note that we have moved the branch for submission. Feel free to let me know if you have any further questions!

**Major Comments:**
1.	**For instance, I believe that requiring users to manually select the B-splines basis, without providing a default option for the B-spline argument, makes the package less "user-friendly" than it could be. Including a default option for the univariate case (and for the bivariate setting, if possible) would be beneficial.**
Response:
We agree with the reviewer that the R package would be more user-friendly if there is a default pre-specified B-spline basis. However, we eventually decided not to incorporate a default specification because B-spline approximation is inherently a data-adaptive approach, and the proper choice of B-spline basis depends on the Phase I and II sample sizes and the number of continuous and discrete covariates (Tao et al. 2021). A pre-specified B-spline basis simply will not work for all scenarios, and we do not want to provide the authors with such a default option that could fail in some scenarios. In the [vignette](https://github.com/dragontaoran/sleev/blob/jossSubmission/vignettes/sleev_vignette.pdf) that comes with the R package, we have provided detailed guidance on how one should construct the B-spline basis under different scenarios and what is the rationale for doing so. We believe that this information will better equip the user with the insight and techniques needed for implementing our methods in a valid and efficient manner.

2.	**In situations where computing k-fold cross-validation is computationally expensive, using a fixed number of B-splines as a starting point could be an effective approach.**
Response:
We totally agree with the reviewer that starting with a (properly chosen) fixed number of B-spline basis functions is computationally efficient and will likely result in valid inference in many scenarios. With that being said, we provide the cross-validation function so that one has the option to choose the optimal B-spline basis in a more rigorous and data-driven manner. In fact, a good routine would be to start with a properly chosen fixed number of B-spline basis function for exploratory analysis. If the results in the exploratory analysis look promising, one may then use cross-validation to either confirm or refine the initial choice of the B-splines.

3.	**Additionally, it would be important to include some relevant methods to enhance usability.**
Response:
Due to the page limit, we are unable to include more details in this manuscript. In the [vignette](https://github.com/dragontaoran/sleev/blob/jossSubmission/vignettes/sleev_vignette.pdf) that comes with the R package, we have included additional data analysis examples with multiple continuous covariates and binary outcomes.

4.	**The term “robust” has been used in the paper. I would appreciate some clarification on what “robust” refers to in this context.**
Response:
Certainly! “Robust” in this paper refers to the fact that the methods implemented in the R package models the error data mechanisms nonparametrically and thus is robust to misspecification of the error models.

5.	**Regarding related packages, are there any other commonly used packages that estimate under these models, aside from the two developed by the author’s colleagues?**
Response:
We are not aware of other methods that can analyze two-phase studies with potentially correlated errors in the outcome and covariate data. Classical measurement error literature largely deals with scenarios where there are measurement errors in the outcome only, covariates only, but not both. In fact, this is a main selling point for our two recent methods publications (Tao et al. 2021; Lotspeich et al. 2022), whose methods are implemented in the R package sleev.

6.	**Additionally, I would suggest including system times in order to compare the different B-spline methods presented in the vignette.**
Response:
Certainly! System time is now included in both the main manuscript and the [vignette](https://github.com/dragontaoran/sleev/blob/jossSubmission/vignettes/sleev_vignette.pdf) that comes with the R package.

**Minor comments:**

1.	**Line 26: The word “and” is missing.**
Thanks for spotting that. The word is now added.

2.	**Lines: 45 and 54. The letter “beta” is missing in the PDF compilation.**
Thanks for noticing this. The letter is now added.

3.	**Line 49: The symbol for independence might not be clear to the reader. I suggest replacing it with the word “independent”.**
Thanks for pointing out this ambiguity. It is now described as “We assume that W, U and \epsilon are independent”.
4.	**After equation (1): There is no mention in the main paper of the probability term that was left out.**
Thanks for pointing this out. We agree that this could be confusing and we added explanation as follows: “….equation (1)…. where $P(\pmb{X^*})$ is left out, because the error-prone covariates are fully observed and thus $P(\pmb{X^*})$ can simply be estimated empirically.”

5.	**Line 62: The equation exceeds the page margins and is unreadable.**
Thanks for noticing this. We now adjusted the equation so that it is within the page margin.

6.	**Equation (2). The variable “s_n” is lowercase but in the sum it appears in uppercase. Additionally, an “i” appears instead of a “k” and some “k”’s are missing.**
Thanks for spotting this. The equation is now updated with the correct letter case for s_n. We did not found discrepancy with “i” and “k” in the formula, but we added parenthesis to avoid potential confusion.

7.	**Lines 115 and 117: The formulas exceed the right margin and can be broken before “degree=3”.**
Thanks for noticing this. We now adjusted the equation so that it is within the page margin.

8.	**Lines 138 and 139: The formulas exceed the right margin as well.**
Thanks for noticing this. We now adjusted the equation so that it is within the page margin.

9.	**Lines 143: The output is missing.**
Thanks for noticing this. The output is now added.

10.	**Table 2: This table is not referenced in the text.**
Thanks for catching this. Table 2 is supposed to referenced at the paragraph where the mock data is introduced. It is now fixed and referenced at row 85.

11.	**Vignette: The reference to Schumaker should be 2007, as this is the third edition (as mentioned in the paper), but it is cited as 1981 here.**
Thanks for pointing out this discrepancy. This reference is now fixed.

12.	**Vignette: Before Equation (1), the regression model is stated as Y=alpha + X * beta + epsilon. However, in the first paragraph, it states alpha + beta^T * X + epsilon. Is X the same in both cases? If so, a transpose might be missing.**
Thanks for pointing out this inconsistency. The linear regression expression are now all changed to alpha + beta^T * X + epsilon.
13.	**Vignette: Equation (3): There are two logarithms in the second sum.**
Thanks for picking up this typo. Equation (3) is now fixed.

References:

Lotspeich, S. C., B. E. Shepherd, G. Amorim, P. A. Shaw, and R. Tao. 2022. “EﬀicientOdds Ratio Estimation Under Two-Phase Sampling Using Error-Prone Data from a Multi-National HIV Research Cohort.”Biometrics78 (4): 1674–85.https://doi.org/10.1111/biom.13512.

Tao, R., S. C. Lotspeich, G. Amorim, P. A. Shaw, and B. E. Shepherd. 2021. “EﬀicientSemiparametric Inference for Two‐phase Studies with Outcome and Covariate MeasurementErrors.”Statistics in Medicine40 (3): 725–38.https://doi.org/10.1002/sim.8799.
