### 2024-12-17 #7320 @alemermartinez
Thanks, @JiangmeiRubyXiong , for addressing my comments. For clarification, I respond to each item below:
    1. I understand your point. My suggestion was more about an “internal” procedure to select the B-spline basis. However, I realize this could require significantly more programming.
    2. and 3. Agreed.
    4. Regarding the explanation of what “robust” means in this context, my concern wasn’t about misunderstanding the concept. Rather, since “robust” can have different interpretations, I believe it would be helpful to explicitly clarify its meaning in the paper, perhaps with a statement similar to the one provided in the vignette.
    5. Ok.
    6. Excellent.
For Equation (2): Based on the result in Line 65, I still believe that w_i​ should be w_k​ in the first term of the two-term formula. Additionally, what do w and u without subscripts represent? Please compare this with Equation (2) in the vignette.
Lastly, I noticed two additional typos:
    • Line 43: There is a missing space between the period and “Suppose.”
    • At the end of Equation (1): There is a period at the end of the line.
