### 2024-12-19 #7320 @JiangmeiRubyXiong
Thanks @alemermartinez for the follow-up comments! Here is our response:

**1. I understand your point. My suggestion was more about an “internal” procedure to select the B-spline basis. However, I realize this could require significantly more programming.**
Thank you, we agree that an “internal” procedure to select the B-spline basis would be ideal. However, as the reviewer pointed out, it could be a nontrivial task and require a significant amount of programming if we would like to incorporate an “fully automatic” B-spline selection mechanism that is versatile enough to accommodate a wide range of scenarios under different number/types of covariates, phase one/two sample sizes, and regression models. While appealing, we consider such a major update to be outside the scope of the current paper/software, whose main goal is to provide for the first time, a user-friendly software that can perform efficient inference in two-phase measurement error studies. We will attempt a fully automatic B-spline selection mechanism in a future update of the package.

**4. Regarding the explanation of what “robust” means in this context, my concern wasn’t about misunderstanding the concept. Rather, since “robust” can have different interpretations, I believe it would be helpful to explicitly clarify its meaning in the paper, perhaps with a statement similar to the one provided in the vignette.**
Thank you for the clarification. We now modified our introduction of the SMLE at line 28 to clarify this:
_"SMLE is an estimator that analyzes two-phase data that combines the error-prone data on all records with the validated data on a subsample. SMLE is highly efficient and robust because it utilizes all available data from both phases while not making any parametric assumption on the error model."_

**For Equation (2): Based on the result in Line 65, I still believe that w_i​ should be w_k​ in the first term of the two-term formula. Additionally, what do w and u without subscripts represent? Please compare this with Equation (2) in the vignette.**
Thank you very much for pointing out this oversight. Equation 2 is now updated and modified with w_k and u_k in place. w, u without subscript is referring to the value of W and U in equation 1. Equation 2 should use w_k and u_k, the distinct observed values of W and U, as defined in the paragraph before equation 2.

**Lastly, I noticed two additional typos:
• Line 43 **(now line 45)**: There is a missing space between the period and “Suppose.”
• At the end of Equation (1): There is a period at the end of the line.**
Thank you for pointing out these typos. They are now fixed in the manuscript.
