### 2024-12-21 #7320 @aalfons
@JiangmeiRubyXiong @jbytecode

First of all, my apologies for taking so long with the review. I do believe that the package implements important functionality that shows some promise, but in its current form the package seems unfinished and not ready for publication. I do emphasize the first part of the previous sentence, and I hope that you find the comments below useful for providing an updated version of the software and manuscript.

Here are my comments on the package:

1. Functions in the package simply return lists. It's best practice for modeling functions to assign a certain class to the results, and then provide methods for `print()` and `summary()` to display the results, as well as methods for `coef()`, `fitted()`, `residuals()`, and `predict()` functions to access or compute the corresponding information from the object. Also some other methods might be relevant, such as `logLik()`. Without adhering to these common design principles, I cannot recommend to accept the submission for publication.

2. Why is there a function to select the B-spline basis via cross-validation only available for the linear model, but not for the logistic model? I would assume that selecting an appropriate B-spline basis via cross-validation is also a relevant for the logistic regression model?

3. Arguments use a mix of different naming conventions: some use underscores for separation (`Y_unval`, `X_unval`, `hn_scale`), others do not (`Bspline`, `noSE`), and while most of them are in lower case, some are in upper case (`TOL`, `MAX_ITER`). I think some consistency in style would increase the usability of the functions. At least I kept getting things wrong when playing around with the package. I know that it is daunting to change the user interface, and there may be legitimate reasons not to do it. On the other hand, the package does not have any reverse dependencies at the moment, so there will never be a better time to implement such changes. For a transition period, you can keep dual arguments for some time (old name and new name), and give a warning when the old name of an argument is used (with a hint that this argument will be removed in the future).

4. The behavior of functions `linear2ph()` and `logistic2ph()` seems different from common design principles for modeling functions in R. For instance, the `coefficients` component returned by these functions seems to contain a matrix of the coefficient estimates, standard errors, test statistics, and p-values. It's common that the model fitting function would only return the coefficient estimates, and that the `summary()` method would provide the inference. There can be legitimate reasons to deviate from this, for instance if the procedures for estimation and inference cannot easily be disentangled in the implementation. But then this difference from common design principles should clearly be addressed in the documentation so that the user knows what to expect. And even in that case, it would be beneficial to the user if the `print()`, `summary()`, and `coefficient()` methods still behave in the usual ways, i.e., similar to those for classes `"lm"` as returned by function `lm()`.

5. The documentation of the returned object is also unclear. For instance, it is not mentioned that the out put is a list (see also my above comment that this should be changed by assigning a class), the documentation jumps straight to describing the list components. Moreover, the description of the list components is also unclear. For instance, the the description of the `coefficient` component simply states "Stores the analysis results." What are the "analysis results"? In what data type or class are they stored? Please provide complete information on the different list components and what data type or class is used to store them.

6. The titles of the help page for funciton `cv_linear2ph()` consists of two sentences and is identical to the **Description** section. Please provide proper (short) titles for all help files and make sure that they are substantially different from the **Description** section.

7. None of the help files seem to contain a **Details** section. I think that providing more details on the estimators in the help files would be beneficial to the user. I do appreciate that there is a vignette and that the references to the papers are given in the help files, but I think that some basic details are also warranted in the help files themselves.

8. I'm going to jump in on a discussion started by @alemermartinez. Thank you for explaining that it may not be possible to come up with a good B-spline basis that works well in all scenarios. However, in the paper you do state that the involvement of B-splines constitutes a difficulty of using the method in practice, but that your package "makes this method readily applicable for practitioners in a user-friendly way". I therefore think that it would be great if the package could provide some utility functions that make the use of B-splines a bit easier on the practitioner. Consider the case study in the manuscript, where you use separate B-spline bases for the two sex groups, which are then combined into one matrix in a specific way and added to the data frame containing the original information. There could be a utility function that does that for the user. For instance, the utility function could take the following arguments: a spline basis for each group, the original data set, and some way to identify which variable in the data set defines the groups. The utility function then returns the data set including the spline bases. Note that this function should work with an arbitrary number of groups, it should work if the data are not necessarily ordered according to the grouping variable, and the data should be returned in the original order of the observations. I can see that applications where different B-spline bases are needed for different groups are quite common, and such a function would greatly enhance usability (in line with your claims on usability) while making the handling of the different B-spline bases less error prone.

Major comments on the manuscript:

1. I appreciate that there has already been a discussion with @alemermartinez regarding clarification the term "robustness". In addition to clarifying the term in the manuscript, I think it would be beneficial if you could include a one- or two-sentence summary on what results you obtained in the referenced papers with respect to robustness. For instance, are there results for consistency under any type of error distribution, or theoretical bounds on the estimation error that hold for any error distribution? Or has robustness been studied empirically by means of simulation? If the latter, which error distributions have been studied? This would provide further details on the robustness of the method without requiring the reader to search for this information the referenced papers.

2. While I understand from the discussion with @alemermartinez that there are no other packages that implement these models, I would be surprised if there are no other packages that are in some form related to what this package are doing. While there may be no packages for such two-phase validated data, there are plenty of packages that address robust or nonparametric estimation of linear and logistic models, and there should also be packages that deal with measurement error/uncertainty in other ways. Hence I think it would be good if you include a paragraph (perhaps as a separate section) on related R packages and how they differ from `sleev`.

3. In the example case study, why do you choose these specific transformations? Most notably, why is CD4 divided by 10 before taking the square root?

4. The manuscript makes claims regarding computational performance compared to packages logreg2ph and TwoPhaseReg. How large is the performance improvement? Can you please show this, either in the paper (if space allows) or otherwise somewhere else (for instance the GitHub README)?

Minor comments on the manuscript:

- Missing space between the two sentences on page 2, line 43.

- Page 3, line 77: "The package can be installed and loaded from the R CRAN library or Github." First, the R in CRAN already stands for the R language, so "R CRAN" is redundant. Second, there is some confusion on terminology. In R, the term "library" refers to a path on a machine that contains installed packages. CRAN is a central repository of user-contributed package, not a library. Third, packages are not loaded from CRAN or GitHub. They can be installed from CRAN or GitHub, but afterwards they are loaded from the local library. So this should be changed to something like "The package can be installed from CRAN or Github." You should also give the link to CRAN, not just to GitHub.

- I think there is another missing space between the two sentences on page 4, line 134. In general, also @alemermartinez found quite a few typos as well as notation and language issues, so please make sure to always carefully proofread and check your work.

- Output from the example is not well formatted and difficult to read (page 4, lines 157-160). Please fix this. (And please use `summary()` to show the results once the requested changes to the package are implemented, see my comment above.)

- Please add a reference for `R`. You can use `citation()` to find out how `R` wants to be cited.

Coding style of example in the manuscript and the GitHub README:

- If one simply copies and pastes the example code from the paper and runs it in `R`, one gets error messages as the command to load package `sleev` is omitted. I suggest to include `load("sleev")` in the example code. Note the quotes around the package name, cf. my next comment.
- It's best practice to use `data("mock.vccc")` instead of `data(mock.vccc)`, as the former highlights that `mock.vccc` is not (yet) an object that exists in the user's workspace.
- Please always use appropriate spacing to make the code readable, for instance around operators or argument assignments, e.g., `sum(mock.vccc$Sex == 0) / n` or
```
Bspline0 <- splines::bs(x = mock.vccc$VL_unval_l10[mock.vccc$Sex == 0], df = sn_0,
                        degree = 3, intercept = TRUE)
```
- Using function `cbind()` in `data <- data.frame(cbind(mock.vccc, Bspline))` is redundant, `data <- data.frame(mock.vccc, Bspline)` achieves the same result.

GitHub:

- It seems that no LICENSE file is included in the GitHub repository, or am I overlooking it?

- In the **Install** section of the README, you only mention how to install the package from GitHub. Please also include information on how to install the package from CRAN.

- Please add community guidelines to the README that explain how third parties can 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Or have I overlooked this information?

All the best with making the necessary adjustments, and please let me know if anything in my comments is unclear.
