### 2025-02-17 #7320 @JiangmeiRubyXiong
@jbytecode Sure! We are applying lots of changes across the package, and therefore we also need to rewrite corresponding parts in the manuscript and vignette. For now, we implemented all changes needed in the package and writing the manuscript & vignette. May I know if there is a deadline for response?
