+++ paper.md
@@ -72,7 +72,7 @@ have low bias and are as efficient as possible.
 
 There are some R packages that are effective for data with measurement errors. The R package `decon` performs nonparametric regression for measurement error models [@wang2011deconvolution]. The R package `simex` and `mecor` approximates the error distribution with different parametric approaches [@lederer2006short; @nab2021mecor]. All of them deals with errors in either outcome or the covariates but not both, and there are no ways to combine the two phases of data with these packages.
 
-SMLE is an estimator that analyzes two-phase data that combines the error-prone data on all records with the validated data on a subsample. SMLE is highly efficient because it utilizes all available data from both phases. [@tao2021efficient; @lotspeich2022efficient]. SMLE is also robust, since it is not making any parametric assumption on the error model. One of the proofs for the SMLE's robustness is a simulation analysis in @tao2021efficient, where SMLE is far less biased with error model misspecification when compared with a moment-based estimator. Moreover, SMLE allows model with error-prone outcome and error-prone covariates. Still, in
+SMLE is an estimator that analyzes two-phase data that combines the error-prone data on all records with the validated data on a subsample. SMLE is highly efficient because it utilizes all available data from both phases. [@tao2021efficient; @lotspeich2022efficient]. SMLE is also robust, since it is not making any parametric assumption on the error model. One of the proofs for the SMLE's robustness is a simulation analysis in @tao2021efficient, where SMLE is far less biased with error model misspecification when compared with a moment-based estimator. Moreover, SMLE allows error-prone outcome and error-prone covariates in the same model. Still, in
 practice these estimators can be difficult to implement, as they involve
 approximating nuisance conditional densities using B-splines [@schumaker2007spline]
 and then maximizing the semiparametric likelihood via a sophisticated EM
