+++ paper.md
@@ -39,21 +39,24 @@ authors:
   - name: Ran Tao
     affiliation: '1,4'
     orcid: 0000-0002-1106-2923
+editor_options: 
+  markdown: 
+    wrap: 72
 ---
 
 # Summary
 
 Data with measurement error in the outcome, covariates, or both are not
 uncommon, particularly with the increased use of routinely collected
-data for biomedical research. In settings with error-prone data, often
-only a subsample of study data is validated, also known as two-phase
-studies. The sieve maximum likelihood estimator (SMLE), which combines
-the error-prone data on all records with the validated data on a
-subsample, is a highly efficient and robust estimator to analyze such
+data for biomedical research. With error-prone data, often only a
+subsample of study data is validated; such setting are known as
+two-phase studies. The sieve maximum likelihood estimator (SMLE), which
+combines the error-prone data on all records with the validated data on
+a subsample, is a highly efficient and robust estimator to analyze such
 data. However, given their complexity, a computationally efficient and
-user-friendly tool is needed to obtain SMLEs. R package `sleev` fills this
-gap by making semiparametric likelihood-based inference using SMLEs for
-error-prone two-phase data in settings with binary and continuous
+user-friendly tool is needed to obtain SMLEs. R package `sleev` fills
+this gap by making semiparametric likelihood-based inference using SMLEs
+for error-prone two-phase data in settings with binary and continuous
 outcomes. Functions from this package can be used to analyze data with
 error-prone responses and covariates.
 
@@ -62,48 +65,75 @@ error-prone responses and covariates.
 Routinely collected data are being used frequently in biomedical
 research, such as electronic health records. However, these data tend to
 be error-prone, and careless use of these data could lead to biased
-estimates and misleading research findings [@duan2016empirical]. To avoid invalid
-study results, trained experts carefully verify and extract data
-elements. However, it is usually only feasible to validate data for a
-subset of records or variables. After validation, researchers have
-error-prone pre-validation data for all records (phase one data) and error-free validated data on a subset of records (phase two data).
-Analyses aims to combine the two types of data to obtain estimates that
+estimates and misleading research findings [@duan2016empirical]. To
+avoid invalid study results, trained experts carefully verify and
+extract data elements. However, it is usually only feasible to validate
+data for a subset of records or variables. After validation, researchers
+have error-prone pre-validation data for all records (phase one data)
+and error-free validated data on a subset of records (phase two data).
+Analyses aim to combine the two types of data to obtain estimates that
 have low bias and are as efficient as possible.
 
-There are some R packages that are effective for data with measurement errors. The R package `decon` performs nonparametric regression for measurement error models [@wang2011deconvolution]. The R package `simex` and `mecor` approximates the error distribution with different parametric approaches [@lederer2006short; @nab2021mecor]. All of them deals with errors in either outcome or the covariates but not both, and there are no ways to combine the two phases of data with these packages.
+There are several [R](https://www.r-project.org/) packages that address
[…]
