+++ R/logistic2ph.R
@@ -24,6 +24,7 @@
 #' `logistic2ph()` returns an object of class `"logistic2ph"`. The function `coef()` is used to obtain the coefficients of the fitted model. The function `summary()` is used to obtain and print a summary of results.
 #'
 #' An object of class `"logistic2ph"` is a list containing at least the following components:
+#' \item{call}{the matched call.}
 #' \item{coefficients}{A named vector of the logistic regression coefficient estimates.}
 #' \item{covariance}{The covariance matrix of the logistic regression coefficient estimates.}
 #' \item{converge}{In parameter estimation, if the EM algorithm converges, then \code{converge = TRUE}. Otherwise, \code{converge = FALSE}.}
@@ -38,6 +39,9 @@
 #' @export
 
 logistic2ph <- function(y_unval = NULL, y = NULL, x_unval = NULL, x = NULL, z = NULL, b_spline = NULL, data = NULL, hn_scale = 1, se = TRUE, tol = 1E-4, max_iter = 1000, verbose = FALSE) {
+  # Store the function call
+  model_call <- match.call()
+
   # variable name change
   Y_unval = y_unval; Y = y ; X_unval = x_unval; X = x; Z = z
   Bspline = b_spline; noSE = !se; TOL = tol; MAX_ITER = max_iter
@@ -249,9 +253,9 @@ logistic2ph <- function(y_unval = NULL, y = NULL, x_unval = NULL, x = NULL, z = 
     gradient_theta <- .calculateGradient(w_t, n, theta_design_mat, comp_dat_all[, Y], muVector)
     hessian_theta <- .calculateHessian(theta_design_mat, w_t, muVector, n, mus_theta);
     new_theta <- tryCatch(expr = prev_theta - (solve(hessian_theta) %*% gradient_theta),
-      error = function(err) {
-        matrix(NA, nrow = nrow(prev_theta))
-        })
+                          error = function(err) {
+                            matrix(NA, nrow = nrow(prev_theta))
+                          })
     if (any(is.na(new_theta))) {
       suppressWarnings(new_theta <- matrix(glm(formula = theta_formula, family = "binomial", data = data.frame(comp_dat_all), weights = w_t)$coefficients, ncol = 1))
       # browser()
@@ -368,44 +372,44 @@ logistic2ph <- function(y_unval = NULL, y = NULL, x_unval = NULL, x = NULL, z = 
 
     ## Calculate pl(theta) -------------------------------------------------
     od_loglik_theta <- observed_data_loglik(N = N,
-      n = n,
-      Y_unval = Y_unval,
-      Y = Y,
-      X_unval = X_unval,
-      X = X,
-      Z = Z,
-      Bspline = Bspline,
-      comp_dat_all = comp_dat_all,
-      theta_pred = theta_pred,
-      gamma_pred = gamma_pred,
-      theta = new_theta,
-      gamma = new_gamma,
-      p = new_p)
+                                            n = n,
+                                            Y_unval = Y_unval,
+                                            Y = Y,
+                                            X_unval = X_unval,
+                                            X = X,
+                                            Z = Z,
+                                            Bspline = Bspline,
+                                            comp_dat_all = comp_dat_all,
+                                            theta_pred = theta_pred,
+                                            gamma_pred = gamma_pred,
+                                            theta = new_theta,
+                                            gamma = new_gamma,
+                                            p = new_p)
 
     I_theta <- matrix(data = od_loglik_theta,
                       nrow = nrow(new_theta),
                       ncol = nrow(new_theta))
 
     single_pert_theta <- sapply(X = seq(1, ncol(I_theta)),
-      FUN = pl_theta,
-      theta = new_theta,
-      h_N = h_N,
-      n = n,
[…]

[…]
