+++ paper.md
@@ -93,7 +93,7 @@ SMLE is highly efficient because it uses all available data from both
 phases [@tao2021efficient; @lotspeich2022efficient]. SMLE is also
 robust, since it does not make any parametric assumptions on the error
 model. For example, @tao2021efficient performed a set of simulations
-highlighting the SMLE\'s robustness to different error mechanisms
+highlighting the SMLE's robustness to different error mechanisms
 including settings where the errors had non-zero mean and were
 multiplicative. Moreover, SMLE allows error-prone outcome and
 error-prone covariates in the same model. Still, in practice these
@@ -106,12 +106,20 @@ user-friendly way. `sleev` integrates and extends R packages `logreg2ph`
 and `TwoPhaseReg`, two primitive R packages developed with the original
 methods papers [@tao2021efficient; @lotspeich2022efficient]. These two
 packages lacked proper documentation and were difficult to use.
-`logreg2ph` was computationally slow. To promote the use of SMLE in
-two-phase data, extensive work has been done to create `sleev`, a
-computationally efficient and user-friendly R package to analyze
-two-phase, error-prone data. Specifically, in `sleev` we rewrote the
-core algorithms of `logreg2ph` in C++ to speed up the computation (Table
-1), and we unified the syntax across functions.
+`logreg2ph` was computationally slow.
+
+To promote the use of SMLE in two-phase data, extensive work has been
+done to create `sleev`, a computationally efficient and user-friendly R
+package to analyze two-phase, error-prone data. Specifically, in `sleev`
+we rewrote the core algorithms of `logreg2ph` in C++ to speed up the
+computation, and we unified the syntax across functions. To compare the
+computational times, we set up simulations with the same code in the
+vignette. The simulations included phase-1 and phase-2 sampe sizes of
+2087 and 835, respectively, and we performed on a *ACCRE information
+that I will include once the code finishes running*. Across 100
+simulations, the previous `logreg2ph` took on averaage xx seconds with a
+standard deviation of yy seconds, while the current `logisitc2ph` in
+`sleev` took xx seconds with a standard deviation of yy seconds.
 
 # SMLE for linear regression
 
@@ -206,15 +214,47 @@ their standard errors. The package can be installed from
 includes two main functions: linear2ph() and logistic2ph(), to fit
 linear and logistic regressions, respectively, under two-phase sampling
 with an error-prone outcome and covariates. The input arguments are
-similar for the two functions and listed in Table 2. From Table 2, we
+similar for the two functions and listed in Table 1. From Table 1, we
 see that in addition to the arguments for error-prone and error-free
 outcome and covariates, the user needs to specify the B-spline matrix
 $B_{j}^{q}\left( \mathbf{X}_{i}^{*} \right)$ to be used in the
 estimation of the error densities.
 
-The `sleev` package included a dataset constructed to mimic data from
+Table 1: Main arguments for `linear2ph()` and `logistic2ph()`
+
+| Argument | Description                                                                                                                           |
+|:----------------------|:-----------------------------------------------|
+| y_unval  | Column name of unvalidated outcome in the input dataset.                                                                              |
+| y        | Column name of validated outcome in the input dataset. NAs in the input will be counted as individuals not selected in phase two.     |
+| x_unval  | Column names of unvalidated covariates in the input dataset.                                                                          |
+| x        | Column names of validated covariates in the input dataset. NAs in the input will be counted as individuals not selected in phase two. |
+| z        | Column names of error-free covariates in the input dataset.                                                                           |
+| b_spline | Column names of the B-spline basis in the input dataset.                                                                              |
+| data     | Dataset                                                                                                                               |
+| hn_scale | Scale of the perturbation constant in the variance estimation via the method of profile likelihood. The default is 1.                 |
[…]
