### 2025-03-31 df190a9
Update table in paper
