### 2025-04-05 503735a
Update paper (and some vignette).
