+++ paper.bib
@@ -69,7 +69,7 @@
   publisher={NIH Public Access},
   DOI={10.1214/20-aoas1343}
 }
-@article{wang2011deconvolution,
+@article{decon,
   title={Deconvolution estimation in measurement error models: the R package decon},
   author={Wang, Xiao-Feng and Wang, Bin},
   journal={Journal of statistical software},
@@ -77,7 +77,7 @@
   pages={1--24},
   year={2011}
 }
-@article{nab2021mecor,
+@article{mecor,
   title={Mecor: An R package for measurement error correction in linear regression models with a continuous outcome},
   author={Nab, Linda and van Smeden, Maarten and Keogh, Ruth H and Groenwold, Rolf HH},
   journal={Computer methods and programs in biomedicine},
@@ -86,7 +86,7 @@
   year={2021},
   publisher={Elsevier}
 }
-@article{lederer2006short,
+@article{SIMEX,
   title={A short introduction to the SIMEX and MCSIMEX},
   author={Lederer, Wolfgang and K{\"u}chenhoff, Helmut},
   journal={The Newsletter of the R Project Volume 6/4, October 2006},
@@ -101,4 +101,71 @@
     address = {Vienna, Austria},
     year = {2024},
     url = {https://www.R-project.org/}
+}
+
+@Article{augSIMEX,
+  title = {R package for analysis of data with mixed measurement error and misclassification in covariates: augSIMEX},
+  author = {Zhang, Qihuang and Yi, Grace Y},
+  journal = {Journal of Statistical Computation and Simulation},
+  year = {2019},
+  pages = {1--23},
+  publisher = {Taylor & Francis},
+}
+
+@Manual{attenuation,
+    title = {attenuation: Correcting for Attenuation Due to Measurement Error},
+    author = {Moss, Jonas},
+    year = {2019},
+    note = {R package version 1.0.0},
+    url = {https://CRAN.R-project.org/package=attenuation},
+  }
+
+@Article{GLSME,
+  author = {Hansen, Thomas F. and Bartoszek, Krzysztof},
+  title = {Interpreting the evolutionary regression: the interplay between observational and biological errors in phylogenetic comparative studies},
+  journal = {Systematic Biology},
+  year = {2012},
+  volume = {61},
+  number = {3},
+  pages = {413--425},
+}
+
+@Manual{meerva,
+  title = {meerva: Analysis of Data with Measurement Error Using a Validation
+Subsample},
+  author = {Kremers, Walter K},
+  year = {2021},
+  note = {R package version 0.2-2},
+  url = {https://CRAN.R-project.org/package=meerva},
+}
+
+@Manual{mmc,
+  title = {mmc: Multivariate Measurement Error Correction},
+  author = {Song, Jaejoon},
+  year = {2015},
+  note = {R package version 0.0.3},
+  url = {https://CRAN.R-project.org/package=mmc},
+}
+
+@Manual{refitME,
+  title = {refitME: Measurement Error Modelling using MCEM},
+  author = {Stoklosa, Jakub and Hwang, Wenhan and Warton, David},
+  year = {2021},
+  note = {R package version 1.2.2},
+  url = {https://CRAN.R-project.org/package=refitME},
+}
+
+@Manual{eivtools,
+    title = {eivtools: Measurement Error Modeling Tools},
+    author = {Lockwood, J.R.},
+    year = {2018},
+    note = {R package version 1.0.8},
+    url = {https://cran.r-project.org/web/packages/eivtools/},
+  }
+
+@book{fuller2009measurement,
+  title={Measurement error models},
+  author={Fuller, Wayne A},
+  year={2009},
+  publisher={John Wiley \& Sons}
 }
