+++ paper.md
@@ -49,181 +49,184 @@ editor_options:
 Data with measurement error in the outcome, covariates, or both are not
 uncommon, particularly with the increased use of routinely collected
 data for biomedical research. With error-prone data, often only a
-subsample of study data is validated; such setting are known as
+subsample of study data is validated; such settings are known as
 two-phase studies. The sieve maximum likelihood estimator (SMLE), which
 combines the error-prone data on all records with the validated data on
-a subsample, is a highly efficient and robust estimator to analyze such
+a subsample, is a highly efficient and robust method to analyze such
 data. However, given their complexity, a computationally efficient and
-user-friendly tool is needed to obtain SMLEs. R package `sleev` fills
-this gap by making semiparametric likelihood-based inference using SMLEs
-for error-prone two-phase data in settings with binary and continuous
-outcomes. Functions from this package can be used to analyze data with
-error-prone responses and covariates.
+user-friendly tool is needed to obtain the SMLEs. The R package `sleev`
+fills this gap by making semiparametric likelihood-based inference using
+the SMLEs for error-prone two-phase data in settings with binary and
+continuous outcomes. Functions from this package can be used to analyze
+data with error-prone binary or continuous responses and error-prone
+covariates.
 
 # Statement of Need
 
 Routinely collected data are being used frequently in biomedical
 research, such as electronic health records. However, these data tend to
-be error-prone, and careless use of these data could lead to biased
-estimates and misleading research findings [@duan2016empirical]. To
-avoid invalid study results, trained experts carefully verify and
-extract data elements. However, it is usually only feasible to validate
-data for a subset of records or variables. After validation, researchers
-have error-prone pre-validation data for all records (phase one data)
-and error-free validated data on a subset of records (phase two data).
-Analyses aim to combine the two types of data to obtain estimates that
-have low bias and are as efficient as possible.
+be error-prone, and using these data without correcting for their
+error-prone nature could lead to biased estimates and misleading
+research findings [@duan2016empirical]. To avoid such invalid study
+results, trained experts carefully verify and extract data elements.
+However, it is usually only feasible to validate data for a subset of
+records or variables. After validation, researchers have error-prone,
+pre-validation data for all records (phase one) and error-free validated
+data on a subset of records (phase two). Analyses aim to combine the two
+types of data to obtain estimates that have low bias and are as robust
+and efficient as possible.
 
 There are several [R](https://www.r-project.org/) packages that address
[…]
