+++ paper.md
@@ -117,11 +117,10 @@ the core algorithms of `logreg2ph` in C++ to speed up the computation,
 and we unified the syntax across functions. To compare the computational
 times, we set up simulations with the same code in the vignette. The
 simulations included phase-one and phase-two sample sizes of 2087 and
-835, respectively, and we performed on a *ACCRE information that I will
-include once the code finishes running*. Across 100 simulations, the
-previous `logreg2ph` took on averaage xx seconds with a standard
-deviation of yy seconds, while the corresponding new function in `sleev`
-took xx seconds with a standard deviation of yy seconds.
+835, respectively, and were performed on a 64-bit Linux OS machine with 8G memory. Across 100 simulations, the
+previous `logreg2ph` took on average 173.19 seconds with a standard
+deviation of 4.40 seconds, while the corresponding new function in `sleev`
+took 109.92 seconds with a standard deviation of 7.91 seconds.
 
 # SMLE for Linear Regression
 
@@ -134,13 +133,12 @@ ${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^{T},\sigma^{2})}^{T}$. When
 we have error-prone data, $Y$ and $\pmb{X}$ are unobserved except for a
 subset of validated records. For unvalidated records (the majority),
 only the error-prone outcome $Y^{*} = Y + W$ and covariates
-$\pmb{X}^{\pmb{*}} = \pmb{X} + \pmb{U}$ are observed in place of $Y$ and
+$\pmb{X}^{*} = \pmb{X} + \pmb{U}$ are observed in place of $Y$ and
 $\pmb{X}$, where $W$ and $\pmb{U}$ are the errors for the outcome and
 covariates, respectively. We assume that $W$ and $\pmb{U}$ are
 independent of $\epsilon$ . With potential errors in our data, a naive
 regression analysis using error-prone variables $Y^{*}$ and
-$\pmb{X}^{\pmb{*}}$ could render misleading results
-[@fuller2009measurement].
+$\pmb{X}^{*}$ could render misleading results [@fuller2009measurement].
 
 We assume that the joint density of the complete data
 $\left( Y^{*},\pmb{X}^{*},W,\pmb{U} \right)$ takes the form
@@ -161,35 +159,35 @@ contributions to the log-likelihood can be obtained by integrating out
 $W$ and $\pmb{U}$.
 
 Let
-$\left( Y_{i}^{*},\pmb{X}_{i}^{\pmb{*}},W_{i},\pmb{U}_{i},V_{i},Y_{i},\pmb{X}_{i} \right)$
+$\left( Y_{i}^{*},\pmb{X}_{i}^{*},W_{i},\pmb{U}_{i},V_{i},Y_{i},\pmb{X}_{i} \right)$
 for $i = 1,\ldots,n$ denote independent and identically distributed
 realizations of
-$\left( Y^{*},\pmb{X}^{\pmb{*}},W,\pmb{U},V,Y,\pmb{X} \right)$ in a
+$\left( Y^{*},\pmb{X}^{*},W,\pmb{U},V,Y,\pmb{X} \right)$ in a
[…]
