+++ paper.md
@@ -78,7 +78,7 @@ and efficient as possible.
 
 There are several [R](https://www.r-project.org/) packages that address
 measurement error [@Rlanguage]. A few examples are `augSIMEX`[@augSIMEX],
-`attentuation`[@attentuation], `decon`[@decon], `eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor], `meerva`[@meerva], `mmc`[@mmc],
+`attenuation`[@attenuation], `decon`[@decon], `eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor], `meerva`[@meerva], `mmc`[@mmc],
 `refitME`[@refitME], and `simex` [@SIMEX]. The various R packages
 reflect many different approaches, such as regression calibration
 [@decon], SIMEX (i.e., simulation-extrapolation) [@SIMEX], and
@@ -91,7 +91,7 @@ error-prone phase-one data and the validated phase-two data.
 The sieve maximum likelihood estimator (SMLE) is an estimator that
 analyzes two-phase data by combining the error-prone data on all records
 with the validated data on a subsample. By leveraging all available
-data, the SMLE operates with high efficiency. [@tao2021efficient;
+data, the SMLE operates with high efficiency [@tao2021efficient;
 @lotspeich2022efficient]. Since it does not make any parametric
 assumptions on the error model, the SMLE is also robust. For example,
 @tao2021efficient performed a set of simulations highlighting the SMLE's
@@ -123,12 +123,12 @@ took 109.92 seconds with a standard deviation of 7.91 seconds.
 
 # SMLE for Linear Regression
 
-In this section, we briefly introduce the the SMLE for linear
+In this section, we briefly introduce the SMLE for linear
 regression. Suppose that we want to fit a standard linear regression
 model for a continuous outcome $Y$ and covariates $\pmb{X}$:
-$Y = \alpha + \boldsymbol{\beta}^{T}\pmb{X} + \epsilon$, where
+$Y = \alpha + \boldsymbol{\beta}^\text{T}\pmb{X} + \epsilon$, where
 $\epsilon\sim N({0,\sigma}^{2})$. Our goal is to obtain estimates of
-${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^{T},\sigma^{2})}^{T}$. When
+${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\text{T},\sigma^{2})}^\text{T}$. When
 we have error-prone data, $Y$ and $\pmb{X}$ are unobserved except for a
 subset of validated records. For unvalidated records (the majority),
 only the error-prone outcome $Y^{*} = Y + W$ and covariates
@@ -144,7 +144,7 @@ $\left( Y^{*},\pmb{X}^{*},W,\pmb{U} \right)$ takes the form
 
 $$P\left( Y^{*},\pmb{X}^{*},\ W,\ \pmb{U} \right) = P\left( Y^{*}|\pmb{X}^{*},\ W,\ \pmb{U} \right)P\left( W,\ \pmb{U|}\pmb{X}^{*}\  \right)P\left( \pmb{X}^{*} \right)$$
 
-$$= P_{\pmb{\theta}}\left( Y|X \right)P\left( W,\ \pmb{U|}\pmb{X}^{*} \right)P\left( \pmb{X}^{*} \right)$$
+$$= P_{\pmb{\theta}}\left( Y|X \right)P\left( W,\ \pmb{U|}\pmb{X}^{*} \right)P\left( \pmb{X}^{*} \right),$$
 
 where $P( \cdot )$ and $P\left( \cdot | \cdot \right)$ denote density
[…]
