+++ paper.bib
@@ -38,7 +38,7 @@
 }
 @article{tao2017efficient,
   title={Efficient semiparametric inference under two-phase sampling, with applications to genetic association studies},
-  author={Tao, Ran and Zeng, Donglin and Lin, Dan-Yu},
+  author={Tao, Ran and Zeng, Donglin and Lin, Dan Yu},
   journal={Journal of the American Statistical Association},
   volume={112},
   number={520},
@@ -71,11 +71,12 @@
 }
 @article{decon,
   title={Deconvolution estimation in measurement error models: the R package decon},
-  author={Wang, Xiao-Feng and Wang, Bin},
+  author={Wang, Xiao Feng and Wang, Bin},
   journal={Journal of statistical software},
   volume={39},
   pages={1--24},
-  year={2011}
+  year={2011},
+  DOI={10.18637/jss.v039.i10}
 }
 @article{mecor,
   title={Mecor: An R package for measurement error correction in linear regression models with a continuous outcome},
@@ -84,16 +85,17 @@
   volume={208},
   pages={106238},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.cmpb.2021.106238}
 }
-@article{SIMEX,
-  title={A short introduction to the SIMEX and MCSIMEX},
-  author={Lederer, Wolfgang and K{\"u}chenhoff, Helmut},
-  journal={The Newsletter of the R Project Volume 6/4, October 2006},
-  volume={6},
-  pages={26},
-  year={2006}
-}
+@Manual{SIMEX,
+    title = {simex: SIMEX- And MCSIMEX-Algorithm for Measurement Error Models},
+    author = {Wolfgang Lederer and Heidi Seibold},
+    year = {2019},
+    note = {R package version 1.8},
+    url = {https://CRAN.R-project.org/package=simex},
+  }
+
 @Manual{Rlanguage,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
@@ -106,10 +108,13 @@
 @Article{augSIMEX,
   title = {R package for analysis of data with mixed measurement error and misclassification in covariates: augSIMEX},
   author = {Zhang, Qihuang and Yi, Grace Y},
-  journal = {Journal of Statistical Computation and Simulation},
-  year = {2019},
-  pages = {1--23},
-  publisher = {Taylor & Francis},
+  journal={Journal of Statistical Computation and Simulation},
+  volume={89},
+  number={12},
+  pages={2293--2315},
+  year={2019},
+  publisher={Taylor \& Francis},
+  DOI={10.1080/00949655.2019.1615911}
 }
 
 @Manual{attenuation,
@@ -128,6 +133,7 @@
   volume = {61},
   number = {3},
   pages = {413--425},
+  DOI = {10.1093/sysbio/syr122}
 }
 
 @Manual{meerva,
@@ -167,5 +173,6 @@ Subsample},
   title={Measurement error models},
   author={Fuller, Wayne A},
   year={2009},
-  publisher={John Wiley \& Sons}
+  publisher={John Wiley \& Sons},
+  DOI={10.1002/9780470316665}
 }
