+++ paper.md
@@ -77,16 +77,18 @@ types of data to obtain estimates that have low bias and are as robust
 and efficient as possible.
 
 There are several [R](https://www.r-project.org/) packages that address
-measurement error [@Rlanguage]. A few examples are `augSIMEX`[@augSIMEX],
-`attenuation`[@attenuation], `decon`[@decon], `eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor], `meerva`[@meerva], `mmc`[@mmc],
-`refitME`[@refitME], and `simex` [@SIMEX]. The various R packages
-reflect many different approaches, such as regression calibration
-[@decon], SIMEX (i.e., simulation-extrapolation) [@SIMEX], and
-moment-based corrections [@mecor], to mention a few. Nearly all of these
-existing R packages deal with errors in either the outcome or
-covariates, but not both, and none of these packages permits
-semiparametric likelihood-based inference that incorporates both the
-error-prone phase-one data and the validated phase-two data.
+measurement error [@Rlanguage]. A few examples are
+`augSIMEX`[@augSIMEX], `attenuation`[@attenuation], `decon`[@decon],
+`eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor],
+`meerva`[@meerva], `mmc`[@mmc], `refitME`[@refitME], and `simex`
+[@SIMEX]. The various R packages reflect many different approaches, such
+as regression calibration [@decon], SIMEX (i.e.,
+simulation-extrapolation) [@SIMEX], and moment-based corrections
+[@mecor], to mention a few. Nearly all of these existing R packages deal
+with errors in either the outcome or covariates, but not both, and none
+of these packages permits semiparametric likelihood-based inference that
+incorporates both the error-prone phase-one data and the validated
+phase-two data.
 
 The sieve maximum likelihood estimator (SMLE) is an estimator that
 analyzes two-phase data by combining the error-prone data on all records
@@ -116,22 +118,23 @@ the core algorithms of `logreg2ph` in C++ to speed up the computation,
 and we unified the syntax across functions. To compare the computational
 times, we set up simulations with the same code in the vignette. The
 simulations included phase-one and phase-two sample sizes of 2087 and
-835, respectively, and were performed on a 64-bit Linux OS machine with 8G memory. Across 100 simulations, the
-previous `logreg2ph` took on average 173.19 seconds with a standard
-deviation of 4.40 seconds, while the corresponding new function in `sleev`
-took 109.92 seconds with a standard deviation of 7.91 seconds.
+835, respectively, and were performed on a 64-bit Linux OS machine with
+8G memory. Across 100 simulations, the previous `logreg2ph` took on
+average 173.19 seconds with a standard deviation of 4.40 seconds, while
+the corresponding new function in `sleev` took 109.92 seconds with a
+standard deviation of 7.91 seconds.
 
 # SMLE for Linear Regression
 
-In this section, we briefly introduce the SMLE for linear
-regression. Suppose that we want to fit a standard linear regression
-model for a continuous outcome $Y$ and covariates $\pmb{X}$:
+In this section, we briefly introduce the SMLE for linear regression.
+Suppose that we want to fit a standard linear regression model for a
+continuous outcome $Y$ and covariates $\pmb{X}$:
 $Y = \alpha + \boldsymbol{\beta}^\text{T}\pmb{X} + \epsilon$, where
[…]
