+++ paper.md
@@ -129,9 +129,9 @@ standard deviation of 7.91 seconds.
 In this section, we briefly introduce the SMLE for linear regression.
 Suppose that we want to fit a standard linear regression model for a
 continuous outcome $Y$ and covariates $\pmb{X}$:
-$Y = \alpha + \boldsymbol{\beta}^\text{T}\pmb{X} + \epsilon$, where
+$Y = \alpha + \boldsymbol{\beta}^\rm{T}\pmb{X} + \epsilon$, where
 $\epsilon\sim N({0,\sigma}^{2})$. Our goal is to obtain estimates of
-${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\text{T},\sigma^{2})}^\text{T}$.
+${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\rm{T},\sigma^{2})}^\rm{T}$.
 When we have error-prone data, $Y$ and $\pmb{X}$ are unobserved except
 for a subset of validated records. For unvalidated records (the
 majority), only the error-prone outcome $Y^{*} = Y + W$ and covariates
@@ -147,7 +147,7 @@ $\left( Y^{*},\pmb{X}^{*},W,\pmb{U} \right)$ takes the form
 
 $$P\left( Y^{*},\pmb{X}^{*},\ W,\ \pmb{U} \right) = P\left( Y^{*}|\pmb{X}^{*},\ W,\ \pmb{U} \right)P\left( W,\ \pmb{U|}\pmb{X}^{*}\  \right)P\left( \pmb{X}^{*} \right)$$
 
-$$= P_{\pmb{\theta}}\left( Y|X \right)P\left( W,\ \pmb{U|}\pmb{X}^{*} \right)P\left( \pmb{X}^{*} \right),$$
+$$= P_{\pmb{\theta}}\left( Y|X \right)P\left( W,\ \pmb{U}|\pmb{X}^{*} \right)P\left( \pmb{X}^{*} \right),$$
 
 where $P( \cdot )$ and $P\left( \cdot | \cdot \right)$ denote density
 and conditional density functions, respectively. Specifically,
@@ -178,9 +178,9 @@ $P\left( W_{i},\pmb{U}_{i}|\pmb{X}_{i}^{*} \right)$ using B-spline
 sieves. Specifically, we approximate
 $P\left( w,\pmb{u}|\pmb{X}_{i}^{*} \right)$ and
 $\log P\left( W_{i},\pmb{U}_{i}|\pmb{X}_{i}^{*} \right)$ by
-$\sum_{k = 1}^{m}\text{I}\left( w = w_{k},\pmb{u} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)p_{kj}$
+$\sum_{k = 1}^{m}\rm{I}\left( w = w_{k},\pmb{u} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)p_{kj}$
 and
-$\sum_{k = 1}^{m}\text{I}\left( W_{i} = w_{k},\pmb{U}_{i} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)\log p_{kj}$,
[…]
