+++ paper.md
@@ -1,6 +1,5 @@
 ---
-title: 'sleev: An R Package for Semiparametric Likelihood Estimation with Errors in
-  Variables'
+title: 'sleev: An R Package for Semiparametric Likelihood Estimation with Errors in Variables'
 tags:
   - R
   - measurement error
@@ -129,9 +128,9 @@ standard deviation of 7.91 seconds.
 In this section, we briefly introduce the SMLE for linear regression.
 Suppose that we want to fit a standard linear regression model for a
 continuous outcome $Y$ and covariates $\pmb{X}$:
-$Y = \alpha + \boldsymbol{\beta}^\rm{T}\pmb{X} + \epsilon$, where
+$Y = \alpha + \boldsymbol{\beta}^\mathrm{T}\pmb{X} + \epsilon$, where
 $\epsilon\sim N({0,\sigma}^{2})$. Our goal is to obtain estimates of
-${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\rm{T},\sigma^{2})}^\rm{T}$.
+${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\mathrm{T},\sigma^{2})}^\mathrm{T}$.
 When we have error-prone data, $Y$ and $\pmb{X}$ are unobserved except
 for a subset of validated records. For unvalidated records (the
 majority), only the error-prone outcome $Y^{*} = Y + W$ and covariates
@@ -178,9 +177,9 @@ $P\left( W_{i},\pmb{U}_{i}|\pmb{X}_{i}^{*} \right)$ using B-spline
 sieves. Specifically, we approximate
 $P\left( w,\pmb{u}|\pmb{X}_{i}^{*} \right)$ and
 $\log P\left( W_{i},\pmb{U}_{i}|\pmb{X}_{i}^{*} \right)$ by
-$\sum_{k = 1}^{m}\rm{I}\left( w = w_{k},\pmb{u} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)p_{kj}$
+$\sum_{k = 1}^{m}\mathrm{I}\left( w = w_{k},\pmb{u} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)p_{kj}$
 and
-$\sum_{k = 1}^{m}\rm{I}\left( W_{i} = w_{k},\pmb{U}_{i} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)\log p_{kj}$,
+$\sum_{k = 1}^{m}\mathrm{I}\left( W_{i} = w_{k},\pmb{U}_{i} = \pmb{u}_{k} \right)\sum_{j = 1}^{s_{n}}B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)\log p_{kj}$,
 respectively. Here,
 $\left\{ \left( w_{1},\pmb{u}_{1} \right),...,\n\left( w_{m},\pmb{u}_{m} \right) \right\}$
 are the $m$ distinct observed $\left( W,\pmb{U} \right)$ values from the
@@ -191,7 +190,7 @@ coefficient associated with $B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)$
[…]
