+++ paper.md
@@ -60,320 +60,3 @@ continuous outcomes. Functions from this package can be used to analyze
 data with error-prone binary or continuous responses and error-prone
 covariates.
 
-# Statement of Need
-
-Routinely collected data are being used frequently in biomedical
-research, such as electronic health records. However, these data tend to
-be error-prone, and using these data without correcting for their
-error-prone nature could lead to biased estimates and misleading
-research findings [@duan2016empirical]. To avoid such invalid study
-results, trained experts carefully verify and extract data elements.
-However, it is usually only feasible to validate data for a subset of
-records or variables. After validation, researchers have error-prone,
-pre-validation data for all records (phase one) and error-free validated
-data on a subset of records (phase two). Analyses aim to combine the two
-types of data to obtain estimates that have low bias and are as robust
-and efficient as possible.
-
-There are several [R](https://www.r-project.org/) packages that address
-measurement error [@Rlanguage]. A few examples are
-`augSIMEX`[@augSIMEX], `attenuation`[@attenuation], `decon`[@decon],
-`eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor],
-`meerva`[@meerva], `mmc`[@mmc], `refitME`[@refitME], and `simex`
-[@SIMEX]. The various R packages reflect many different approaches, such
-as regression calibration [@decon], SIMEX (i.e.,
-simulation-extrapolation) [@SIMEX], and moment-based corrections
-[@mecor], to mention a few. Nearly all of these existing R packages deal
-with errors in either the outcome or covariates, but not both, and none
-of these packages permits semiparametric likelihood-based inference that
-incorporates both the error-prone phase-one data and the validated
-phase-two data.
-
-The sieve maximum likelihood estimator (SMLE) is an estimator that
-analyzes two-phase data by combining the error-prone data on all records
-with the validated data on a subsample. By leveraging all available
-data, the SMLE operates with high efficiency [@tao2021efficient;
-@lotspeich2022efficient]. Since it does not make any parametric
-assumptions on the error model, the SMLE is also robust. For example,
-@tao2021efficient performed a set of simulations highlighting the SMLE's
-robustness to different error mechanisms including settings where the
-errors had non-zero mean and were multiplicative. Moreover, the SMLE
-allows error-prone outcome and error-prone covariates in the same model.
-Still, in practice these estimators can be difficult to implement, as
-they involve approximating nuisance conditional densities using
-B-splines [@schumaker2007spline] and then maximizing the semiparametric
-likelihood via a sophisticated EM algorithm [@tao2017efficient]. Here,
-we present the R package `sleev`, which makes the SMLE readily
-applicable for practitioners in a user-friendly way. `sleev` integrates
-and extends primitive R packages, `logreg2ph` and `TwoPhaseReg`,
-developed with the original methods papers [@tao2021efficient;
-@lotspeich2022efficient]. These two packages lacked proper documentation
-and were difficult to use. `logreg2ph` was also computationally slow.
-
-To promote the use of the SMLE, extensive work has been done to create
[…]
