+++ paper.md
@@ -122,3 +122,86 @@ simulations included phase-one and phase-two sample sizes of 2087 and
 average 173.19 seconds with a standard deviation of 4.40 seconds, while
 the corresponding new function in `sleev` took 109.92 seconds with a
 standard deviation of 7.91 seconds.
+
+
+# SMLE for Linear Regression
+
+In this section, we briefly introduce the SMLE for linear regression.
+Suppose that we want to fit a standard linear regression model for a
+continuous outcome $Y$ and covariates $\pmb{X}$:
+$Y = \alpha + \boldsymbol{\beta}^\mathrm{T}\pmb{X} + \epsilon$, where
+$\epsilon\sim N({0,\sigma}^{2})$. Our goal is to obtain estimates of
+${\pmb{\theta} = (\alpha,\boldsymbol{\beta}^\mathrm{T},\sigma^{2})}^\mathrm{T}$.
+When we have error-prone data, $Y$ and $\pmb{X}$ are unobserved except
+for a subset of validated records. For unvalidated records (the
+majority), only the error-prone outcome $Y^{*} = Y + W$ and covariates
+$\pmb{X}^{*} = \pmb{X} + \pmb{U}$ are observed in place of $Y$ and
+$\pmb{X}$, where $W$ and $\pmb{U}$ are the errors for the outcome and
+covariates, respectively. We assume that $W$ and $\pmb{U}$ are
+independent of $\epsilon$ . With potential errors in our data, a naive
+regression analysis using error-prone variables $Y^{*}$ and
+$\pmb{X}^{*}$ could render misleading results [@fuller2009measurement].
+
+We assume that the joint density of the complete data
+$\left( Y^{*},\pmb{X}^{*},W,\pmb{U} \right)$ takes the form
+
+$$P\left( Y^{*},\pmb{X}^{*},\ W,\ \pmb{U} \right) = P\left( Y^{*}|\pmb{X}^{*},\ W,\ \pmb{U} \right)P\left( W,\ \pmb{U|}\pmb{X}^{*}\  \right)P\left( \pmb{X}^{*} \right)$$
+
+$$= P_{\pmb{\theta}}\left( Y|X \right)P\left( W,\ \pmb{U}|\pmb{X}^{*} \right)P\left( \pmb{X}^{*} \right),$$
+
+where $P( \cdot )$ and $P\left( \cdot | \cdot \right)$ denote density
+and conditional density functions, respectively. Specifically,
+$P_{\pmb{\theta}}\left( Y|\pmb{X} \right)$ then refers to the
+conditional density function of the linear regression model of $Y$ and
+$\pmb{X}$. Denote the validation indicator variable by $V$, with $V = 1$
+indicating that a record was validated and $V = 0$ otherwise. For
+records with $V = 0$, their measurement errors
+$\left( W,\pmb{U} \right)$ are missing, and therefore their
+contributions to the log-likelihood can be obtained by integrating out
+$W$ and $\pmb{U}$.
+
+Let
+$\left( Y_{i}^{*},\pmb{X}_{i}^{*},W_{i},\pmb{U}_{i},V_{i},Y_{i},\pmb{X}_{i} \right)$
+for $i = 1,\ldots,n$ denote independent and identically distributed
[…]
