+++ paper.md
@@ -205,3 +205,175 @@ The SMLEs for logistic regression are similar to linear regression and
 described in the [package vignette](https://github.com/dragontaoran/sleev/blob/main/vignettes/sleev_vignette.pdf),
 and the theoretical properties can be found in @lotspeich2022efficient.
 
+# Functionalities of the `sleev` R Package
+
+The `sleev` package provides a user-friendly way to obtain the SMLEs and
+their standard errors. The package can be installed from
+[CRAN](https://cran.r-project.org/web/packages/sleev/index.html) or
+[GitHub](https://github.com/dragontaoran/sleev). The `sleev` package
+includes two main functions: `linear2ph()` and `logistic2ph()`, to fit
+linear and logistic regressions, respectively, under two-phase sampling
+with an error-prone outcome and covariates. The input arguments are
+similar for the two functions and listed in Table 1. In addition to the
+arguments for error-prone and error-free outcome and covariates, the
+user needs to specify the B-spline matrix
+$B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)$ to be used in the estimation
+of the error densities.
+
+Table 1: Main arguments for the `linear2ph()` and `logistic2ph()`
+functions
+
+| Argument | Description|
+|:---------|:--------------------------------------------------|
+| y_unval  | Column name of unvalidated outcome in the input dataset. |
+| y        | Column name of validated outcome in the input dataset. `NA`s in the input will be counted as individuals not selected in phase two.|
+| x_unval  | Column names of unvalidated covariates in the input dataset. |
+| x        | Column names of validated covariates in the input dataset. `NA`s in the input will be counted as individuals not selected in phase two. |
+| z        | Column names of error-free covariates in the input dataset.                                                                             |
+| b_spline | Column names of the B-spline basis in the input dataset.                                                                                |
+| data     | Dataset                                                                                                                                 |
+| hn_scale | Scale of the perturbation constant in the variance estimation via the method of profile likelihood. The default is `1`.                 |
+| se       | Standard errors of the parameter estimators will be estimated when set to TRUE. The default is `TRUE`.                                  |
+| tol      | Convergence criterion. The default is `0.0001`.                                                                                         |
+| max_iter | Maximum number of iterations in the EM algorithm. The default is `1000`.                                                                |
+| verbose  | Print analysis details when set to `TRUE`. The default is `FALSE`.                                                                      |
+
+# Example: Case study with mock data
+
+For demonstration, the `sleev` package includes a dataset constructed to
+mimic data from the Vanderbilt Comprehensive Care Clinic (VCCC) patient
+records from @giganti2020accounting. Table 2 describes the variables in
+this dataset.
+
+Table 2: Data dictionary for `mock.vccc`
+
+| Name      | Status      | Type       | Description                                         |
+|:-----------------|:------------|:------------|:----------------------------|
+| ID        | error-free  |            | Patient ID                                          |
+| VL_unval  | error-prone | continuous | Viral load (VL) at antiretroviral therapy (ART)     |
+| VL_val    | validated   | continuous | initiation                                          |
[…]
