+++ paper.md
@@ -77,10 +77,9 @@ and efficient as possible.
 
 There are several [R](https://www.r-project.org/) packages that address
 measurement error [@Rlanguage]. A few examples are
-`augSIMEX`[@augSIMEX], `attenuation`[@attenuation], `decon`[@decon],
-`eivtools`[@eivtools], `GLSME`[@GLSME], `mecor`[@mecor],
-`meerva`[@meerva], `mmc`[@mmc], `refitME`[@refitME], and `simex`
-[@SIMEX]. The various R packages reflect many different approaches, such
+`augSIMEX` [@augSIMEX], `attenuation` [@attenuation], `decon` [@decon],
+`eivtools` [@eivtools], `GLSME` [@GLSME], `mecor` [@mecor],
+`meerva` [@meerva], `mmc` [@mmc], `refitME` [@refitME], and `simex` [@SIMEX]. The various R packages reflect many different approaches, such
 as regression calibration [@decon], SIMEX (i.e.,
                                            simulation-extrapolation) [@SIMEX], and moment-based corrections
 [@mecor], to mention a few. Nearly all of these existing R packages deal
@@ -97,7 +96,7 @@ data, the SMLE operates with high efficiency [@tao2021efficient;
 assumptions on the error model, the SMLE is also robust. For example,
 @tao2021efficient performed a set of simulations highlighting the SMLE's
 robustness to different error mechanisms including settings where the
-errors had non-zero mean and were multiplicative. Moreover, the SMLE
+errors had non-zero mean or were multiplicative. Moreover, the SMLE
 allows error-prone outcome and error-prone covariates in the same model.
 Still, in practice these estimators can be difficult to implement, as
 they involve approximating nuisance conditional densities using
@@ -115,13 +114,13 @@ To promote the use of the SMLE, extensive work has been done to create
 analyze two-phase, error-prone data. Specifically, in `sleev` we rewrote
 the core algorithms of `logreg2ph` in C++ to speed up the computation,
 and we unified the syntax across functions. To compare the computational
-times, we set up simulations with the same code in the vignette. The
+times, we set up simulations with the same code in the [vignette](https://github.com/dragontaoran/sleev/blob/main/vignettes/sleev_vignette.pdf). The
 simulations included phase-one and phase-two sample sizes of 2087 and
 835, respectively, and were performed on a 64-bit Linux OS machine with
-8G memory. Across 100 simulations, the previous `logreg2ph` took on
-average 173.19 seconds with a standard deviation of 4.40 seconds, while
-the corresponding new function in `sleev` took 109.92 seconds with a
-standard deviation of 7.91 seconds.
+8G memory. Across 100 simulations, the previous `logreg2ph` took an
+average of 289.44 seconds with a standard deviation of 8.83 seconds to perform the analysis, while
+the corresponding new function in `sleev` only took an average of 122.32 seconds with a
+standard deviation of 8.18 seconds.
 
 
 # SMLE for Linear Regression
@@ -145,7 +144,7 @@ $\pmb{X}^{*}$ could render misleading results [@fuller2009measurement].
 We assume that the joint density of the complete data
 $\left( Y^{*},\pmb{X}^{*},W,\pmb{U} \right)$ takes the form
 
-$$P\left( Y^{*},\pmb{X}^{*},\ W,\ \pmb{U} \right) = P\left( Y^{*}|\pmb{X}^{*},\ W,\ \pmb{U} \right)P\left( W,\ \pmb{U|}\pmb{X}^{*}\  \right)P\left( \pmb{X}^{*} \right)$$
[…]
