### 2025-04-11 a00f033
Fix table width attempt 1
