+++ paper.bib
@@ -1,0 +1,178 @@
+@inproceedings{duan2016empirical,
+  title={An empirical study for impacts of measurement errors on EHR based association studies},
+  author={Duan, Rui and Cao, Ming and Wu, Yonghui and Huang, Jing and Denny, Joshua C and Xu, Hua and Chen, Yong},
+  booktitle={AMIA Annual Symposium Proceedings},
+  volume={2016},
+  pages={1764},
+  year={2016},
+  organization={American Medical Informatics Association}
+}
+@article{tao2021efficient,
+  title={Efficient semiparametric inference for two-phase studies with outcome and covariate measurement errors},
+  author={Tao, Ran and Lotspeich, Sarah C and Amorim, Gustavo and Shaw, Pamela A and Shepherd, Bryan E},
+  journal={Statistics in medicine},
+  volume={40},
+  number={3},
+  pages={725--738},
+  year={2021},
+  publisher={Wiley Online Library},
+  DOI={10.1002/sim.8799}
+}
+@article{lotspeich2022efficient,
+  title={Efficient odds ratio estimation under two-phase sampling using error-prone data from a multi-national HIV research cohort},
+  author={Lotspeich, Sarah C and Shepherd, Bryan E and Amorim, Gustavo GC and Shaw, Pamela A and Tao, Ran},
+  journal={Biometrics},
+  volume={78},
+  number={4},
+  pages={1674--1685},
+  year={2022},
+  publisher={Oxford University Press},
+  DOI={10.1111/biom.13512}
+}
+@book{schumaker2007spline,
+  title={Spline functions: basic theory},
+  author={Schumaker, Larry},
+  year={2007},
+  publisher={Cambridge university press},
+  DOI={10.1017/CBO9780511618994}
+}
+@article{tao2017efficient,
+  title={Efficient semiparametric inference under two-phase sampling, with applications to genetic association studies},
+  author={Tao, Ran and Zeng, Donglin and Lin, Dan Yu},
+  journal={Journal of the American Statistical Association},
+  volume={112},
+  number={520},
+  pages={1468--1476},
+  year={2017},
+  publisher={Taylor \& Francis},
+  DOI={10.1080/01621459.2017.1295864}
+}
+@article{murphy2000profile,
+  title={On profile likelihood},
+  author={Murphy, Susan A and Van der Vaart, Aad W},
+  journal={Journal of the American Statistical Association},
+  volume={95},
+  number={450},
+  pages={449--465},
+  year={2000},
+  publisher={Taylor \& Francis},
+  DOI={10.2307/2669386}
+}
+@article{giganti2020accounting,
+  title={Accounting for dependent errors in predictors and time-to-event outcomes using electronic health records, validation samples, and multiple imputation},
+  author={Giganti, Mark J and Shaw, Pamela A and Chen, Guanhua and Bebawy, Sally S and Turner, Megan M and Sterling, Timothy R and Shepherd, Bryan E},
+  journal={The annals of applied statistics},
+  volume={14},
+  number={2},
+  pages={1045},
+  year={2020},
+  publisher={NIH Public Access},
+  DOI={10.1214/20-aoas1343}
+}
+@article{decon,
+  title={Deconvolution estimation in measurement error models: the R package decon},
+  author={Wang, Xiao Feng and Wang, Bin},
+  journal={Journal of statistical software},
+  volume={39},
+  pages={1--24},
+  year={2011},
+  DOI={10.18637/jss.v039.i10}
+}
+@article{mecor,
+  title={Mecor: An R package for measurement error correction in linear regression models with a continuous outcome},
[…]
