### 2025-04-24 #7320 @JiangmeiRubyXiong
@jbytecode @aalfons Thank you very much for your patience. We applied major updates regarding the package (both on Github and CRAN), the manuscript and the vignette per reviewer comment. Here are our detailed response regarding the comments:

**Comments on the package:**

1. _Functions in the package simply return lists. It's best practice for modeling functions to assign a certain class to the results, and then provide methods for `print()` and `summary()` to display the results, as well as methods for `coef()`, `fitted()`, `residuals()`, and `predict()` functions to access or compute the corresponding information from the object. Also some other methods might be relevant, such as `logLik()`. Without adhering to these common design principles, I cannot recommend to accept the submission for publication._

Thank you for your suggestion, and we agree that common design principles are important. We made the following changes to the package:

- Function `linear2ph()` now returns list object with class `linear2ph` and function `logistic2ph()` returns list object with class `logistic2ph`. The structure and output of these two classes are aligned for consistency.
- Functions `print()`, `coef()`, `coefficients()`, and `summary()` are added for both classes `linear2ph` and `logistic2ph`. The details of these functions are described in the response to comment 4.
- The other suggested commands – `fitted()`, `residuals()`, and `predict()` – are not well-defined yet for the outcome and covariate measurement errors under two-phase designs. Therefore, they are not implemented. In fact, the existing literature on two-phase studies with measurement errors has been largely focusing on study designs and inference procedures. Literature on model diagnostics is lacking. It is not straightforward to calculate fitted values and residuals for study subjects whose true outcome and covariate values are unknown. We will incorporate these features into the package once new methods in these areas become available.

2. _Why is there a function to select the B-spline basis via cross-validation only available for the linear model, but not for the logistic model? I would assume that selecting an appropriate B-spline basis via cross-validation is also relevant for the logistic regression model?_

Thank you for catching this discrepancy. Function `cv_logistic2ph()` is now added to the package. The arguments and output of this function are in line with `cv_linear2ph()`’s.

3. _Arguments use a mix of different naming conventions: some use underscores for separation (`Y_unval`, `X_unval`, `hn_scale`), others do not (`Bspline`, `noSE`), and while most of them are in lower case, some are in upper case (`TOL`, `MAX_ITER`). I think some consistency in style would increase the usability of the functions. At least I kept getting things wrong when playing around with the package. I know that it is daunting to change the user interface, and there may be legitimate reasons not to do it. On the other hand, the package does not have any reverse dependencies at the moment, so there will never be a better time to implement such changes. For a transition period, you can keep dual arguments for some time (old name and new name) and give a warning when the old name of an argument is used (with a hint that this argument will be removed in the future)._

Thank you for noticing this, and we agree that the argument names should be consistent. We made the following changes to the argument names:
|Before | Y_unval | Y | X_unval | X | Z | Bspline | noSE | TOL | MAX_ITER|
|-- | -- | -- | -- | -- | -- | -- | -- | -- | --|
|__After__| y_unval | y | x_unval | x | z | b_spline | se | tol | max_iter|

4. _The behavior of functions `linear2ph()` and `logistic2ph()` seems different from common design principles for modeling functions in R. For instance, the coefficients component returned by these functions seems to contain a matrix of the coefficient estimates, standard errors, test statistics, and p-values. It's common that the model fitting function would only return the coefficient estimates, and that the `summary()` method would provide the inference. There can be legitimate reasons to deviate from this, for instance if the procedures for estimation and inference cannot easily be disentangled in the implementation. But then this difference from common design principles should clearly be addressed in the documentation so that the user knows what to expect. And even in that case, it would be beneficial to the user if the `print()`, `summary()`, and `coefficient()` methods still behave in the usual ways, i.e., similar to those for the class `"lm"` as returned by function `lm()`._

We agree with the reviewer and the made the changes accordingly so that `print()` gives the model formula, algorithm convergence status, and coefficients; `coefficient()` gives the coefficients only; and `summary()` gives the matrix of coefficient estimates, standard errors, test statistics, and p-values.

5. _The documentation of the returned object is also unclear. For instance, it is not mentioned that the output is a list (see also my above comment that this should be changed by assigning a class), the documentation jumps straight to describing the list components. Moreover, the description of the list components is also unclear. For instance, the description of the coefficient component simply states, "Stores the analysis results." What are the "analysis results"? In what data type or class are they stored? Please provide complete information on the different list components and what data type or class is used to store them._
6. _Pease provide proper (short) titles for all help files and make sure that they are substantially different from the Description section._
7. _None of the help files seem to contain a Details section. I think that providing more details on the estimators in the help files would be beneficial to the user. I do appreciate that there is a vignette and that the references to the papers are given in the help files, but I think that some basic details are also warranted in the help files themselves._

Responses to comments 5–7:
Thank you for observing this, and we agree that our documentation can be improved a lot. We now updated documentation for all functions as follows:

- Title is revised to be appropriate length and reflects the content.
- Added/updated function description, class (if any) and output description.
- We modified the description of the arguments and output to be more informative. For example, the coefficient argument is now “A named vector of the linear regression coefficient estimates.”

8. _I'm going to jump in on a discussion started by @alemermartinez . Thank you for explaining that it may not be possible to come up with a good B-spline basis that works well in all scenarios. However, in the paper you do state that the involvement of B-splines constitutes a difficulty of using the method in practice, but that your package "makes this method readily applicable for practitioners in a user-friendly way". I therefore think that it would be great if the package could provide some utility functions that make the use of B-splines a bit easier on the practitioner. Consider the case study in the manuscript, where you use separate B-spline bases for the two sex groups, which are then combined into one matrix in a specific way and added to the data frame containing the original information. There could be a utility function that does that for the user. For instance, the utility function could take the following arguments: a spline basis for each group, the original data set, and some way to identify which variable in the data set defines the groups. The utility function then returns the data set including the spline bases. Note that this function should work with an arbitrary number of groups, it should work if the data are not necessarily ordered according to the grouping variable, and the data should be returned in the original order of the observations. I can see that applications where different B-spline bases are needed for different groups are quite common, and such a function would greatly enhance usability (in line with your claims on usability) while making the handling of the different B-spline bases less error prone._

Thank you for the constructive suggestion! We agree that this is a good midpoint where the usability is improved without forcing an inflexible default on users. We now added a function `spline2ph()` that creates B-splines basis matrix and combines it with the data input so that the output is readily usable for analysis with `linear2ph()` and `logistic2ph()`. The function can take an arbitrary number of variables and allows for stratified B-splines in arbitrary number of groups. The corresponding sections on the manuscript and package vignette are modified to reflect this change.

**Major comments on the manuscript:**

1. _I appreciate that there has already been a discussion with [@alemermartinez](https://github.com/alemermartinez) regarding clarification on the term "robustness". In addition to clarifying the term in the manuscript, I think it would be beneficial if you could include a one- or two-sentence summary on what results you obtained in the referenced papers with respect to robustness. For instance, are there results for consistency under any type of error distribution, or theoretical bounds on the estimation error that hold for any error distribution? Or has robustness been studied empirically by means of simulation? If the latter, which error distributions have been studied? This would provide further details on the robustness of the method without requiring the reader to search for this information in the referenced papers._

Yes, we agree. We now modified the text and added a brief description in the third paragraph (line 40) of the statement of need section to demonstrate this, as follows:

By leveraging all available data, the SMLE operates with high efficiency (Lotspeich et al., 2022; Tao et al., 2021). Since it does not make any parametric assumptions on the error model, the SMLE is also robust. For example, Tao et al. (2021) performed a set of simulations highlighting the SMLE’s robustness to different error mechanisms including settings where the errors had non-zero mean or were multiplicative.

2. _While I understand from the discussion with @alemermartinez that there are no other packages that implement these models, I would be surprised if there are no other packages that are in some form related to what this package are doing. While there may be no packages for such two-phase validated data, there are plenty of packages that address robust or nonparametric estimation of linear and logistic models, and there should also be packages that deal with measurement error/uncertainty in other ways. Hence, I think it would be good if you include a paragraph (perhaps as a separate section) on related R packages and how they differ from sleev._

Yes, this is a very good point. As our package is mostly focused on measurement error data, we added a paragraph describing other R packages that handles data with measurement error (line 29), as follows:

There are several R packages that address measurement error (R Core Team, 2024), including augSIMEX (Zhang & Yi, 2019), attenuation (Moss, 2019), decon (Wang & Wang, 2011), eivtools (Lockwood, 2018), GLSME (Hansen & Bartoszek, 2012), mecor (Nab et al., 2021), meerva (Kremers, 2021), mmc (Song, 2015), refitME (Stoklosa et al., 2021), and simex (Lederer & Seibold, 2019). The various R packages reflect many different approaches, such as regression calibration (Wang & Wang, 2011), SIMEX (i.e., simulation-extrapolation) (Lederer & Seibold, 2019), and moment-based corrections (Nab et al., 2021), to mention a few. Nearly all of these existing R packages deal with errors in either the outcome or covariates, but not both, and none of these packages permits efficient inference that incorporates both the error-prone phase-one data and the validated phase-two data.

3. _In the example case study, why do you choose these specific transformations? Most notably, why is CD4 divided by 10 before taking the square root?_

This example is consistent with Lotspeich et al., 2022, where CD4 was “rescaled to units of 10 cells per microliter.” Although dividing by 10 prior to taking the square root is not necessary, it was done in an attempt to improve interpretation and to pull the estimated coefficients away from zero because a 1-unit change on this new scale constitutes a clinically meaningful (and less incremental) change in CD4.

4. _The manuscript makes claims regarding computational performance compared to packages logreg2ph and TwoPhaseReg. How large is the performance improvement? Can you please show this, either in the paper (if space allows) or otherwise somewhere else (for instance the GitHub README)?_

Thank you for catching this. We need to correct our statement as the package does not significantly improve computational efficiency over TwoPhaseReg, which was already quite fast. However, our package is significantly faster than logreg2ph. We included this information in the last paragraph of the “Statement of Need” section (line 55), as follows:

Specifically, in sleev we rewrote the core algorithms of logreg2ph in C++ to speed up the computation, and we unified the syntax across functions. To compare the computational times, we set up simulations with the same code in the package vignette. The simulations included phase-one and phase-two sample sizes of 2087 and 835, respectively, and were performed on a 64-bit Linux OS machine with 8G memory. Across 100 simulations, the previous logreg2ph took an average of 289.44 seconds with a standard deviation of 8.83 seconds to perform the analysis, while the corresponding new function in sleev only took an average of 122.32 seconds with a standard deviation of 8.18 seconds.

**Minor comments on the manuscript:**

1. _Missing space between the two sentences on page 2, line 43._

Thank you for catching this, it is now fixed.

2. _Page 3, line 77: "The package can be installed and loaded from the R CRAN library or GitHub." First, the R in CRAN already stands for the R language, so "R CRAN" is redundant. Second, there is some confusion on terminology. In R, the term "library" refers to a path on a machine that contains installed packages. CRAN is a central repository of user-contributed package, not a library. Third, packages are not loaded from CRAN or GitHub. They can be installed from CRAN or GitHub, but afterwards they are loaded from the local library. So, this should be changed to something like "The package can be installed from CRAN or GitHub." You should also give the link to CRAN, not just to GitHub._

Thank you for catching this, and we agree that we should use better language. The sentence is now changed to “The package can be installed from CRAN or GitHub.” (line 100), and the CRAN link is now included.

3. _I think there is another missing space between the two sentences on page 4, line 134. In general, also [@alemermartinez](https://github.com/alemermartinez) found quite a few typos as well as notation and language issues, so please make sure to always carefully proofread and check your work._

Thank you for catching this typo, and we have also carefully proofread the submitted version.

4. _Output from the example is not well formatted and difficult to read (page 4, lines 157-160). Please fix this. (And please use summary() to show the results once the requested changes to the package are implemented, see my comment above.)_

Thank you for bringing our attention to this issue. `summary()` is now used to display the result with reasonable outlay.

5. _Please add a reference for R. You can use citation() to find out how R wants to be cited._

Thank you for catching this. The reference for R is now added in line 29.

**Coding style of example in the manuscript and the GitHub README:**

1. _If one simply copies and pastes the example code from the paper and runs it in R, one gets error messages as the command to load package sleev is omitted. I suggest including library("sleev") in the example code. Note the quotes around the package name, cf. my next comment._

Yes, this is a very good point. library("sleev") is now added to the example code in line 121.

2. _It's best practice to use data("mock.vccc") instead of data(mock.vccc), as the former highlights that mock.vccc is not (yet) an object that exists in the user's workspace._

Thank you for catching this. data("mock.vccc") is now used in the manuscript instead of data(mock.vccc) (line 122).

3. _Please always use appropriate spacing to make the code readable, for instance around operators or argument assignments, e.g., `sum(mock.vccc$Sex == 0) / n` or
`Bspline0 <- splines::bs(x = mock.vccc$VL_unval_l10[mock.vccc$Sex == 0], df = sn_0, degree = 3, intercept = TRUE)`_

Thank you for your suggestion on code writing conventions. All example code in the manuscript and vignette is modified accordingly.

4. Using function `cbind()` in `data <- data.frame(cbind(mock.vccc, Bspline))` is redundant, `data <- data.frame(mock.vccc, Bspline)` achieves the same result.

Thank you and we agree this is a redundancy in our code. Though the changes made above to create a splines function in our package meant we no longer needed this line of code.

**GitHub:**

1. It seems that no LICENSE file is included in the GitHub repository, or am I overlooking it?

Thank you for bringing this to our attention. The reason we did not included the license file initially is that `LICENSE.txt` on top level conflicts with CRAN submission. We now included a `LICENSE_sleev.txt` file in folder `./inst`.

2. In the Install section of the README, you only mention how to install the package from GitHub. Please also include information on how to install the package from CRAN.

We added the following in the README file:

or install the CRAN version of the package by running:
```
install.packages("sleev")
```

3. Please add community guidelines to the README that explain how third parties can 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Or have I overlooked this information?

Thank you for bringing this to our attention. We now added community guideline to the README as follows:

If you wish to make suggestions, report a problem, or need more support with package use, please email Dr. Ran Tao (r DOT tao AT vumc DOT org) or Dr. Sarah Lotspeich (lotspes AT wfu DOT edu).
