### 2025-06-18 #7320 @JiangmeiRubyXiong
Hi @aalfons ! We updated package and please look at everything together. Here are the breakdown of the changes and responses:

Major comments:

- Perhaps you overlooked to instruct the editorialbot to generate the latest version?
  - Yes, thank you for pointing it out. I just instructed the editorialbot to regenerate the last version, right below this message. Please review the latest version, which responds to all comments regarding code and writing.
- I think it would be useful if function `spline2ph` generates a reasonable default for the variable names of the B-spline basis. For ease of usability, this should not be a required argument that causes an error if missing.
  - Yes, we agree that it would improve the usability of this package. We now modified function `spline2ph()` so that it does not require input of B-spline basis variable names anymore.
- Most package help files do not contain any examples. The standard for well documented packages is that each help file has at least some example. The examples from the paper could serve as the basis for the help files.
  - We completely agree that including example code in the help file would significantly enhance the usability of the package. However, we’ve encountered some challenges in doing so.  The SMLE nonparametrically approximates the error distributions, which is desirable because it is robust to model misspecification. Nevertheless, it is inherently computationally demanding (even with the C++ codes). We previously included examples, but they took more than a few seconds to run, which led to complaints from CRAN. As a result, we had to remove them. If you have any suggestions for how we might work around this limitation, we would be more than happy to incorporate them.
- In the README.md on https://github.com/dragontaoran/sleev (which is the GitHub link given in the paper), I do not see the sentence on community guidelines that you mention in your response above. Am I overlooking something?
  - Thank you for bringing this to our attention. The guideline is now added to README.

Minor comments:

- Make sure that you consistently use = or <- to assign objects in the code examples. Most of the time, you use <- but on line 123 you use =.
  - Thank you for pointing out this inconsistency. We sifted through both manuscript and vignette and made sure we only use `<-` consistently.
- The vignette title not properly registered and shown as "Your Vignette Title" on the CRAN page of the package, see _link_
  - Thank you for pointing this out. We now modified this vignette title to be "sleev_vignette", see [here](https://cran.r-project.org/web/packages/sleev/index.html).
- Note that there is a way to keep the LICENSE in the usual place in the root directory, all one needs to do is to tell R CMD build not to include it when building the package. This allows GitHub users to find the LICENSE in the expected place. For this purpose, you can create a file .Rbuildignore in the root directory of the package, where you list all files that should not be included when building the package. See one of my packages on https://www.github.com/aalfons/robustHD (also published in JOSS) for an example on how to do this.
  - We appreciate your information on this issue and it is very helpful. We now have LICENCE in the top level of folder and visible in GitHub.
