### 2025-06-23 #7320 @aalfons
@JiangmeiRubyXiong - Thank you for the changes and for the detailed response. So far I have only read your response, but I didn't yet have another look at the paper or package. I first have some suggestions based on your response alone.

- Regarding the examples: A package really is not very useful if it doesn't have examples. If the take some time to run, there are things you can do. For instance, you can wrap them in a \donttest{} statement - this has the advantage that those examples can still be chacked automatically with `R CMD check --run-donttest` (see the Writing R Extensions manual). If CRAN still complains, you can comment out the example code (and add a comment why), which is better than removing the examples altogether.
- "sleev_vignette" is not a very informative title to be displayed on the CRAN page. Please register the actual title of the vignette so that a user has an indication of its content.

I'll wait with checking the paper and software again until examples are added to the help files, as I find this a crucial point.
