### 2025-07-01 #7320 @JiangmeiRubyXiong
Hi @aalfons, thank you for the information! Examples are now added to the help files. The vignette title is now updated to "sleev: Semiparametric Likelihood Estimation with Errors in Variables" on the CRAN page.
