### 2025-07-29 #7320 @aalfons
@JiangmeiRubyXiong I now went over the manuscript, package, and your responses from April 25 and June 18. Thanks for all the efforts and detailed responses!

I'm happy with everything, except that the reference to R seems misplaced in the text on line 29: "There are several R packages that address measurement error (R Core Team, 2024),  including [...]" Please change this to something like "There are several packages for R (R Core Team, 2024) that address measurement error, including [...]"

@jbytecode I already checked off all the boxes in my reviewer checklist.
