### 2025-07-29 d3c7950
Update paper.md R reference.
