### 2025-07-29 cea98df
Update paper.md typo fix
