### 2025-07-29 0fcf51c
Update paper.bib DOI
