+++ paper.bib
@@ -94,6 +94,7 @@
     year = {2019},
     note = {R package version 1.8},
     url = {https://CRAN.R-project.org/package=simex},
+    DOI = {10.32614/CRAN.package.simex}
   }
 
 @Manual{Rlanguage,
@@ -102,7 +103,8 @@
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
     year = {2024},
-    url = {https://www.R-project.org/}
+    url = {https://www.R-project.org/},
+    DOI = {10.32614/r.manuals}
 }
 
 @Article{augSIMEX,
@@ -123,6 +125,7 @@
     year = {2019},
     note = {R package version 1.0.0},
     url = {https://CRAN.R-project.org/package=attenuation},
+    DOI = {10.32614/CRAN.package.attenuation}
   }
 
 @Article{GLSME,
@@ -143,6 +146,7 @@ Subsample},
   year = {2021},
   note = {R package version 0.2-2},
   url = {https://CRAN.R-project.org/package=meerva},
+DOI = {10.32614/CRAN.package.meerva}
 }
 
 @Manual{mmc,
@@ -151,6 +155,7 @@ Subsample},
   year = {2015},
   note = {R package version 0.0.3},
   url = {https://CRAN.R-project.org/package=mmc},
+  DOI = {10.32614/CRAN.package.mmc}
 }
 
 @Manual{refitME,
@@ -159,6 +164,7 @@ Subsample},
   year = {2021},
   note = {R package version 1.2.2},
   url = {https://CRAN.R-project.org/package=refitME},
+  DOI = {10.32614/CRAN.package.refitME}
 }
 
 @Manual{eivtools,
@@ -167,6 +173,7 @@ Subsample},
     year = {2018},
     note = {R package version 1.0.8},
     url = {https://cran.r-project.org/web/packages/eivtools/},
+    DOI = {10.32614/CRAN.package.eivtools}
   }
 
 @book{fuller2009measurement,
