### 2025-07-30 b3175c4
fix capitalizations in bibtex
