+++ paper.bib
@@ -33,7 +33,7 @@
   title={Spline functions: basic theory},
   author={Schumaker, Larry},
   year={2007},
-  publisher={Cambridge university press},
+  publisher={Cambridge University Press},
   DOI={10.1017/CBO9780511618994}
 }
 @article{tao2017efficient,
@@ -70,7 +70,7 @@
   DOI={10.1214/20-aoas1343}
 }
 @article{decon,
-  title={Deconvolution estimation in measurement error models: the R package decon},
+  title={Deconvolution estimation in measurement error models: the {R} package decon},
   author={Wang, Xiao Feng and Wang, Bin},
   journal={Journal of statistical software},
   volume={39},
@@ -79,7 +79,7 @@
   DOI={10.18637/jss.v039.i10}
 }
 @article{mecor,
-  title={Mecor: An R package for measurement error correction in linear regression models with a continuous outcome},
+  title={Mecor: An {R} package for measurement error correction in linear regression models with a continuous outcome},
   author={Nab, Linda and van Smeden, Maarten and Keogh, Ruth H and Groenwold, Rolf HH},
   journal={Computer methods and programs in biomedicine},
   volume={208},
@@ -140,7 +140,7 @@
 }
 
 @Manual{meerva,
-  title = {meerva: Analysis of Data with Measurement Error Using a Validation
+  title = {{meerva}: Analysis of Data with Measurement Error Using a Validation
 Subsample},
   author = {Kremers, Walter K},
   year = {2021},
@@ -150,7 +150,7 @@ DOI = {10.32614/CRAN.package.meerva}
 }
 
 @Manual{mmc,
-  title = {mmc: Multivariate Measurement Error Correction},
+  title = {{mmc}: Multivariate Measurement Error Correction},
   author = {Song, Jaejoon},
   year = {2015},
   note = {R package version 0.0.3},
@@ -168,7 +168,7 @@ DOI = {10.32614/CRAN.package.meerva}
 }
 
 @Manual{eivtools,
-    title = {eivtools: Measurement Error Modeling Tools},
+    title = {{eivtools}: Measurement Error Modeling Tools},
     author = {Lockwood, J.R.},
     year = {2018},
     note = {R package version 1.0.8},
