### 2025-07-30 45fbb78
minor updates in paper
