### 2025-07-30 e5659a0
format table headers as H3
