+++ paper.md
@@ -217,7 +217,7 @@ user needs to specify the B-spline matrix
 $B_{j}^{q}\left( \pmb{X}_{i}^{*} \right)$ to be used in the estimation
 of the error densities.
 
-Table 1: Main arguments for the `linear2ph()` and `logistic2ph()`
+### Table 1: Main arguments for the `linear2ph()` and `logistic2ph()`
 functions
 
 | Argument | Description|
@@ -241,7 +241,7 @@ mimic data from the Vanderbilt Comprehensive Care Clinic (VCCC) patient
 records from @giganti2020accounting. Table 2 describes the variables in
 this dataset.
 
-Table 2: Data dictionary for `mock.vccc`
+### Table 2: Data dictionary for `mock.vccc`
 
 | Name        | Status      | Type       | Description                                         |
 |:------------|:------------|:-----------|:----------------------------------------------------|
