### 2025-07-30 #7320 @jbytecode
@JiangmeiRubyXiong - After the first pull request I've just seen the header of the first table is not rendered as expected. This PR fixes this. Please merge. Thank you.
