### 2025-07-30 #7320 @JiangmeiRubyXiong
@jbytecode merged. Thank you!
