### 2025-07-31 b96dc89
Update paper.md minor typo
