### 2025-07-31 #7320 @JiangmeiRubyXiong
Hi @jbytecode! The software version is v1.1.4 and the doi is https://doi.org/10.5281/zenodo.16622343
Let me know if it works!
