### 2025-08-01 #7320 @jbytecode
@JiangmeiRubyXiong -

- In the Zenodo archive the title, the author list, and author metadata (ORCID) should match those in the paper.

In its current form the title is `dragontaoran/sleev: sleev v1.1.4 (JOSS published version)` which should be set to the one of the paper.

Thank you in advance.
