### 2025-08-01 #7320 @JiangmeiRubyXiong
@jbytecode fixed!
