### 2025-08-01 #7320 @jbytecode
@JiangmeiRubyXiong - The title in the Zenodo archive is still

`dragontaoran/sleev: sleev v1.1.4 (JOSS published version)`

The title should be the same with the title of the manuscript.
